% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{asinhtGml2_breaks}
\alias{asinhtGml2_breaks}
\title{Inverse hyperbolic sine breaks (GatingML 2.0 version)}
\usage{
asinhtGml2_breaks(n = 6, equal.space = FALSE, ...)
}
\arguments{
\item{n}{desired number of breaks (the actual number will be different depending on the data range)}

\item{equal.space}{whether breaks at equal-spaced intervals}

\item{...}{parameters passed to asinh.Gml2}
}
\value{
a function generates fasinh or fsinh space
}
\description{
Used to construct \code{\link{asinhtGml2_trans}} object
}
\examples{
data <- 1:1e3
brks.func <- asinhtGml2_breaks(equal.space = TRUE)
brks <- brks.func(data)
brks # fasinh space displayed at raw data scale

#transform it to verify it is equal-spaced at transformed scale
trans.func <- asinh.Gml2()
brks.trans <- trans.func(brks)
brks.trans
}

