% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{asinhtGml2_trans}
\alias{asinhtGml2_trans}
\title{Inverse hyperbolic sine transformation.}
\usage{
asinhtGml2_trans(..., n = 6, equal.space = FALSE)
}
\arguments{
\item{...}{parameters passed to asinh_Gml2}

\item{n}{desired number of breaks (the actual number will be different depending on the data range)}

\item{equal.space}{whether breaks at equal-spaced intervals}
}
\value{
asinhtGml2 transformation object
}
\description{
Used to construct inverse hyperbolic sine transform object.
}
\examples{
asinhtGml2_trans()
}

