\name{booleanFilter}
\alias{booleanFilter}
\title{Constructor for booleanFilter}
\usage{
  booleanFilter(expr, ...,
    filterId = "defaultBooleanFilter")
}
\arguments{
  \item{expr}{\code{expression}}

  \item{...}{further arguments to the expression}

  \item{filterId}{\code{character} identifier}
}
\description{
  Constructor from an expression
}

