% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{cf_backend_type}
\alias{cf_backend_type}
\title{return the cytoframe backend storage format}
\usage{
cf_backend_type(cf)
}
\arguments{
\item{cf}{cytoframe}
}
\value{
one of "mem","h5", "tile"
}
\description{
return the cytoframe backend storage format
}
