% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{cf_get_h5_file_path}
\alias{cf_get_h5_file_path}
\title{return the file path of underlying h5 file
For the in-memory version of cytoframe, it returns empty string.Thus can be used to check whether it is on-disk format.}
\usage{
cf_get_h5_file_path(cf)
}
\arguments{
\item{cf}{cytoframe object}
}
\description{
return the file path of underlying h5 file
For the in-memory version of cytoframe, it returns empty string.Thus can be used to check whether it is on-disk format.
}
