% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{cf_get_uri}
\alias{cf_get_uri}
\alias{cf_get_h5_file_path}
\title{Return the file path of the underlying h5 file}
\usage{
cf_get_uri(cf)

cf_get_h5_file_path(cf)
}
\arguments{
\item{cf}{cytoframe object}
}
\description{
Return the file path of the underlying h5 file
}
\details{
For the in-memory version of cytoframe, it returns an empty string. This can be used to check whether it is on-disk format.
}
\seealso{
Other cytoframe/cytoset IO functions: 
\code{\link{cf_write_disk}()},
\code{\link{cf_write_h5}()},
\code{\link{cs_get_uri}()},
\code{\link{load_cytoframe_from_fcs}()},
\code{\link{load_cytoframe}()},
\code{\link{load_cytoset_from_fcs}()}
}
\concept{cytoframe/cytoset IO functions}
