% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{cf_write_h5}
\alias{cf_write_h5}
\title{save the cytoframe as h5 format}
\usage{
cf_write_h5(cf, filename)
}
\arguments{
\item{cf}{cytoframe object}

\item{filename}{the full path of the output h5 file}
}
\description{
save the cytoframe as h5 format
}
