\name{char2booleanFilter}
\alias{char2booleanFilter}
\title{Constructor for booleanFilter}
\usage{
  char2booleanFilter(expr, ...,
    filterId = "defaultBooleanFilter")
}
\arguments{
  \item{expr}{a \code{character} string}

  \item{...}{further arguments to the expression}

  \item{filterId}{\code{character} identifier}
}
\description{
  Constructor from a character string
}

