% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_GatingSet.R
\name{checkRedundantNodes}
\alias{checkRedundantNodes}
\title{try to determine the redundant terminal(or leaf) nodes that can be removed}
\usage{
checkRedundantNodes(x, path = "auto", ...)
}
\arguments{
\item{x}{\code{GatingSet} or \code{list} of groups(each group is a list of 'GatingSet`). When it is a list, it is usually the outcome from \link{groupByTree}.}

\item{path}{argumented passed to \link{getNodes}. The default value is "auto".}

\item{...}{other arguments passed to \link{getNodes}.}
}
\value{
a list of the character vectors inicating the nodes that are considered to be redundant for each group of GatingSets.
}
\description{
THese leaf nodes make the gating trees to be different from one another and can be removed by the subsequent convevient call 
\link{dropRedundantNodes}.
}
\examples{
\dontrun{
gslist <- list(gs1, gs2, gs3, gs4, gs5)
gs_groups <- groupByTree(gslist)
toRm <- checkRedundantNodes(gs_groups)
}
}
