% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\name{compare.counts}
\alias{compare.counts}
\title{compare the counts to cytobank's exported csv so that the parsing result can be verified.}
\usage{
compare.counts(gs, file, id.vars = c("FCS Filename", "population"))
}
\arguments{
\item{gs}{parsed GatingSet}

\item{file}{the stats file (contains the populatio counts) exported from cytobank.}

\item{id.vars}{either "population" or "FCS filename" that tells whether the stats file format is one population per row or FCS file per row.}
}
\value{
a data.table (in long format) that contains the counts from openCyto and Cytobank side by side.
}
\description{
compare the counts to cytobank's exported csv so that the parsing result can be verified.
}
\examples{
\dontrun{
xmlfile <- system.file("extdata/cytotrol_tcell_cytobank.xml", package = "flowWorkspace")
fcsFiles <- list.files(pattern = "CytoTrol", system.file("extdata", package = "flowWorkspaceData"), full = T)
gs <- parse.gatingML(xmlfile, fcsFiles)
## verify the stats are correct
statsfile <- system.file("extdata/cytotrol_tcell_cytobank_counts.csv", package = "flowWorkspace")
dt_merged <- compare.counts(gs, statsfile, id.vars = "population")
all.equal(dt_merged[, count.x], dt_merged[, count.y], tol = 5e-4)
}
}

