% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\docType{methods}
\name{compensate,cytoframe,matrix-method}
\alias{compensate,cytoframe,matrix-method}
\title{compensate the data in place}
\usage{
\S4method{compensate}{cytoframe,matrix}(x, spillover)
}
\description{
compensate the data in place
}
