% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphGML-methods.R
\docType{methods}
\name{compensate,flowSet,graphGML-method}
\alias{compensate,flowSet,graphGML-method}
\title{compensate a flowSet based on the compensation information stored in graphGML object}
\usage{
\S4method{compensate}{flowSet,graphGML}(x, spillover, ...)
}
\arguments{
\item{x}{flowSet}

\item{spillover}{graphGML}

\item{...}{unused.}
}
\value{
compensated flowSet
}
\description{
compensate a flowSet based on the compensation information stored in graphGML object
}

