% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphGML-methods.R
\docType{methods}
\name{compensate,GatingSet,graphGML-method}
\alias{compensate,GatingSet,graphGML-method}
\title{compensate a GatingSet based on the compensation information stored in graphGML object}
\usage{
\S4method{compensate}{GatingSet,graphGML}(x, spillover, ...)
}
\arguments{
\item{x}{GatingSet}

\item{spillover}{graphGML}

\item{...}{unused.}
}
\value{
compensated GatingSet
}
\description{
compensate a GatingSet based on the compensation information stored in graphGML object
}

