% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet-Methods.R, R/GatingSetList-Methods.R
\docType{methods}
\name{compensate,GatingSet,compensation-method}
\alias{compensate,GatingSet,compensation-method}
\alias{compensate,GatingSetList,compensation-method}
\title{compensate the flow data asssociated with the GatingSet}
\usage{
\S4method{compensate}{GatingSet,compensation}(x, spillover)

\S4method{compensate}{GatingSetList,compensation}(x, spillover)
}
\arguments{
\item{x}{\code{GatingSet} or \code{GatingSetList}}

\item{spillover}{\code{compensation} object}
}
\value{
a \code{GatingSet} or \code{GatingSetList} object with the underling flow data compensated.
}
\description{
The compensation is saved in the GatingSet and can be retrieved by \link{getCompensationMatrices}.
}
\examples{
\dontrun{

cfile <- system.file("extdata","compdata","compmatrix", package="flowCore")
comp.mat <- read.table(cfile, header=TRUE, skip=2, check.names = FALSE)
## create a compensation object
comp <- compensation(comp.mat,compensationId="comp1")
#add it to GatingSet
gs <- compensate(gs, comp)
}
}

