% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{compute_timestep}
\alias{compute_timestep}
\title{compute time step from fcs keyword}
\usage{
compute_timestep(kw, unit.range, timestep.source = c("TIMESTEP", "BTIM"))
}
\arguments{
\item{kw}{list of keywords}

\item{unit.range}{the actual measured time unit range}

\item{timestep.source}{either "TIMESTEP" or "BTIM". prefer to $TIMESTEP keyword when it is non NULL}
}
\description{
compute time step from fcs keyword
}
