% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\name{constructTree}
\alias{constructTree}
\title{Reconstruct the population tree from the GateSets}
\usage{
constructTree(flowEnv, gateInfo)
}
\arguments{
\item{flowEnv}{the enivornment contains the elements parsed by read.gatingML function}

\item{gateInfo}{the data.frame contains the gate name, fcs filename parsed by parse.gateInfo function}
}
\value{
a graphNEL represent the population tree. The gate and population name are stored as nodeData in each node.
}
\description{
Reconstruct the population tree from the GateSets
}

