% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gs.R
\name{convert_backend}
\alias{convert_backend}
\title{convert h5 based gs archive to tiledb}
\usage{
convert_backend(gs_dir, output_dir)
}
\arguments{
\item{gs_dir}{existing gs archive path}

\item{output_dir}{the new gs path}
}
\description{
convert h5 based gs archive to tiledb
}
