% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gs.R
\name{convert_legacy_gs}
\alias{convert_legacy_gs}
\alias{convert_legacy_gslist}
\title{convert the legacy(mixed with R and C++ files) GatingSet archive to the new format(C++ only)}
\usage{
convert_legacy_gs(from, to)

convert_legacy_gslist(from, to)
}
\arguments{
\item{from}{the old archive path}

\item{to}{the new archive path}
}
\description{
Note that it is likely some of the keyword values (mainly offsets e.g. BEGINDATA) may change slightly after the converting due to the rewriting data
to FCS files through write.FCS.
}
\examples{
\dontrun{
convert_legacy_gs(old_gs_path, new_gs_path)
}
}
