% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoset.R
\name{cs_add_cytoframe}
\alias{cs_add_cytoframe}
\title{Add a cytoframe to a cytoset}
\usage{
cs_add_cytoframe(cs, sn, cf)
}
\arguments{
\item{cs}{cytoset}

\item{sn}{sample name to be added}

\item{cf}{cytoframe to be added}
}
\description{
Add a cytoframe to a cytoset
}
