% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoset.R
\name{cs_get_h5_file_path}
\alias{cs_get_h5_file_path}
\title{Return the file path of the underlying h5 files}
\usage{
cs_get_h5_file_path(x)
}
\description{
Return the file path of the underlying h5 files
}
\seealso{
Other cytoframe/cytoset IO functions: \code{\link{cf_get_h5_file_path}},
  \code{\link{cf_write_h5}},
  \code{\link{load_cytoframe_from_fcs}},
  \code{\link{load_cytoframe_from_h5}},
  \code{\link{load_cytoset_from_fcs}}
}
\concept{cytoframe/cytoset IO functions}
