% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoset.R
\name{cs_set_cytoframe}
\alias{cs_set_cytoframe}
\title{update a cytoframe in a cytoset}
\usage{
cs_set_cytoframe(cs, sn, cf)
}
\arguments{
\item{cs}{cytoset}

\item{sn}{sample name}

\item{cf}{cytoframe}
}
\description{
update a cytoframe in a cytoset
}
