% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R, R/cytoset.R
\docType{methods}
\name{cyto_flow_coerce_methods}
\alias{cyto_flow_coerce_methods}
\alias{cytoframe_to_flowFrame}
\alias{flowFrame_to_cytoframe}
\alias{cytoset_to_flowSet}
\alias{flowSet_to_cytoset}
\title{Methods for conversions between cytoframe/cytoset and flowFrame/flowSet}
\usage{
cytoframe_to_flowFrame(fr)

flowFrame_to_cytoframe(fr, ...)

cytoset_to_flowSet(cs)

flowSet_to_cytoset(fs, path = tempfile())
}
\description{
These methods consist of a pair of methods to coerce a \code{cytoframe}
to or from a \code{flowFrame} and another pair to coerce a \code{cytoset}
to or from a \code{flowSet}.
}
\details{
The conversion between the two sets of data container classes mostly entails
a conversion of the back-end representation of the data. \code{cytoframe}
and \code{cytoset} objects contain \code{flowFrame} and \code{flowSet} objects
respectively, so coercion of a \code{cytoframe} to \code{flowFrame} entails
moving the data from the 'C'-level data structure to the corresponding
\code{exprs}, \code{description}, and \code{parameters} slots. Coercion of
a \code{flowFrame} to a \code{cytoframe} entails creation of the 'C'-level
data structure from the \code{flowFrame} slots. The names of each of the
methods are pretty self-explanatory.
}
\section{Methods}{


\describe{
\item{cytoframe_to_flowFrame(object = "cytoframe")}{Returns a \code{flowFrame} object
coerced from a \code{cytoframe} object.}

\item{flowFrame_to_cytoframe(object = "flowFrame")}{Returns a \code{cytoframe} object
coerced from a \code{flowFrame} object.}

\item{cytoset_to_flowSet(object = "cytoset")}{Returns a \code{flowSet} object
coerced from a \code{cytoset} object.}

\item{flowSet_to_cytoset(object = "flowSet")}{Returns a \code{cytoset} object
coerced from a \code{flowSet} object.}

}
}

\examples{
data("GvHD")
fs <- GvHD[1]
cs <- flowSet_to_cytoset(fs)
cf <- cs[[1, returnType="cytoframe"]]
ff <- cytoframe_to_flowFrame(cf)

}
\keyword{methods}
