% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{cyto_stats}
\alias{cyto_stats}
\title{constructor of cyto_stats}
\usage{
cyto_stats(type, channel = "", ancestor = "", percent = "")
}
\arguments{
\item{type}{character stats type. type 'cyto_stats_supported()' for all the supported stats types}

\item{channel}{character channel info}

\item{ancestor}{character ancestor population . only valid for 'Freqof' stats}

\item{percent}{numeric only valid for 'Percentile' stats}
}
\description{
constructor of cyto_stats
}
