% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoctx.R
\name{cytoctx}
\alias{cytoctx}
\alias{print.cytoctx}
\alias{ctx_to_list}
\title{Create cyto context that carries tiledb context parameters}
\usage{
cytoctx(cred = NULL, num_threads = 1L)

print.cytoctx(x, ...)

ctx_to_list(x)
}
\arguments{
\item{cred}{credentials for s3 access. It is a list containing elements of "AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY", "AWS_REGION"
when NULL, read the default credential file from disk (e.g., ~/.aws/credentials)}

\item{num_threads}{tiledb multithread parameters}

\item{x}{cytoctx}
}
\value{
a cytoctx object that is a wrapper around the external pointer,thus it is copy-by-reference object
}
\description{
Create cyto context that carries tiledb context parameters

Convert cytoctx to a list
}
