% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R, R/cytoset.R
\name{cytoframe-labels}
\alias{cytoframe-labels}
\alias{cf_swap_colnames}
\alias{cf_rename_channel}
\alias{cf_rename_marker}
\alias{cs_swap_colnames}
\title{Methods to change channel and marker names for \code{cytoframe} and \code{cytoset} objects}
\usage{
cf_swap_colnames(x, col1, col2)

cf_rename_channel(x, old, new)

cf_rename_marker(x, old, new)

cs_swap_colnames(x, col1, col2)
}
\arguments{
\item{x}{a \code{cytoframe}}

\item{col1}{first channel name to swap}

\item{col2}{second channel name to swap}

\item{old}{old channel or marker name to be changed}

\item{new}{new channel or marker name after change}
}
\description{
The methods allow direct alteration of channel names or marker names
of \code{\link{cytoframe}} and \code{\link{cytoset}} objects. These objects are accessed
by reference and changed in place, so there is no need to assign the return
value of these methods.
}
