% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{get_default_backend}
\alias{get_default_backend}
\alias{set_default_backend}
\title{get/set the default backend format of cytoframe}
\usage{
get_default_backend()

set_default_backend(backend = c("h5", "mem", "tile"))
}
\arguments{
\item{backend}{one of c("h5", "mem",  "tile")}
}
\description{
get/set the default backend format of cytoframe
}
