% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gs.R
\name{delete_gs}
\alias{delete_gs}
\title{delete the archive of GatingSet}
\usage{
delete_gs(path)
}
\arguments{
\item{path}{either a local path or s3 path (e.g. "s3://bucketname/gs_path)}
}
\description{
delete the archive of GatingSet
}
