% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\name{extend}
\alias{extend}
\alias{extend.ellipsoidGate}
\alias{extend.polygonGate}
\alias{extend.rectangleGate}
\title{extend the gate to the minimum and maximum limit of both dimensions
based on the bounding information.
It is equivalent to the behavior of shifting the off-scale boundary events into the gate boundary that is describled in
bounding transformation section of gatingML standard.
The advantage of extending gates instead of shifting data are two folds:
1. Avoid the extra computation each time applying or plotting the gates
2. Avoid changing the data distribution caused by adding the gates}
\usage{
extend(gate, bound, data.range = NULL, plot = FALSE)

\method{extend}{polygonGate}(gate, bound, data.range = NULL, plot = FALSE)

\method{extend}{rectangleGate}(gate, ...)

\method{extend}{ellipsoidGate}(gate, ...)
}
\arguments{
\item{gate}{a flowCore filter/gate}

\item{bound}{numeric matrix representing the bouding information parsed from gatingML. Each row corresponds to a channel.
rownames should be the channel names. colnames should be c("min", "max")}

\item{data.range}{numeric matrix specifying the data limits of each channel. It is used to set the extended value of vertices and must has the same structure as 'bound'.
when it is not supplied, c(-.Machine$integer.max, - .Machine$integer.max) is used.}

\item{plot}{whether to plot the extended polygon.}

\item{...}{other arguments}
}
\description{
extend the gate to the minimum and maximum limit of both dimensions
based on the bounding information.
It is equivalent to the behavior of shifting the off-scale boundary events into the gate boundary that is describled in
bounding transformation section of gatingML standard.
The advantage of extending gates instead of shifting data are two folds:
1. Avoid the extra computation each time applying or plotting the gates
2. Avoid changing the data distribution caused by adding the gates
}

