% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R
\name{extract_cluster_pop_name_from_node}
\alias{extract_cluster_pop_name_from_node}
\title{Extract the population name from the node path
It strips the parent path and cluster method name.}
\usage{
extract_cluster_pop_name_from_node(node, cluster_method_name)
}
\arguments{
\item{node}{population node path}

\item{cluster_method_name}{the name of the clustering method}
}
\description{
Extract the population name from the node path
It strips the parent path and cluster method name.
}
\examples{
extract_cluster_pop_name_from_node("cd3/flowClust_pop1", "flowClust")
#returns "pop1"
}
