% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterObject_Methods.R
\name{filter_to_list}
\alias{filter_to_list}
\alias{filterObject}
\alias{filterObject,default-method}
\alias{filter_to_list,rectangleGate-method}
\alias{filter_to_list,polygonGate-method}
\alias{filter_to_list,quadGate-method}
\alias{filter_to_list,booleanFilter-method}
\alias{filter_to_list,ellipsoidGate-method}
\alias{filter_to_list,logical-method}
\title{convert flowCore filter to a list 
 
It convert the flowCore gate to a list whose structure can be understood by 
underlying c++ data structure.}
\usage{
filter_to_list(x)
}
\arguments{
\item{x}{\code{filter} a flowCore gate. Currently supported gates are:
"rectangleGate", "polygonGate","ellipsoidGate" and "booleanFilter"}
}
\value{
a \code{list}
}
\description{
convert flowCore filter to a list 
 
It convert the flowCore gate to a list whose structure can be understood by 
underlying c++ data structure.
}
