% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{fix_channel_slash}
\alias{fix_channel_slash}
\title{toggle the channel names between '/' and '_' character}
\usage{
fix_channel_slash(chnls, slash_loc = NULL)
}
\arguments{
\item{chnls}{the channel names}

\item{slash_loc}{a list that records the locations of the original slash character within each channel name
so that when restoring slash it won't tamper the the original '_' character.}
}
\value{
the toggled channel names
}
\description{
FlowJoX tends to replace '/' in the original channel names with '_' in gates and transformations.
We need to do the same to the flow data but also need to change it back during the process since
the channel names of the flowSet can't be modified until the data is fully compensated.
}
