% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{flow_trans}
\alias{flow_trans}
\title{helper function to generate a trans objects
Used by other specific trans constructor}
\usage{
flow_trans(name, trans.fun, inverse.fun, equal.space = FALSE, n = 6)
}
\arguments{
\item{name}{transformation name}

\item{trans.fun}{the transform function (only needed when equal.space is TRUE)}

\item{inverse.fun}{the inverse function (only needed when equal.space is TRUE)}

\item{equal.space}{whether breaks at equal-spaced intervals}

\item{n}{desired number of breaks (the actual number will be different depending on the data range)}
}
\description{
helper function to generate a trans objects
Used by other specific trans constructor
}
