% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet-Methods.R
\docType{methods}
\name{flowData,GatingSet-method}
\alias{flowData}
\alias{flowData,GatingSet-method}
\alias{flowData<-}
\alias{flowData<-,GatingSet-method}
\title{Fetch or replace the flowData object associated with a GatingSet .}
\usage{
\S4method{flowData}{GatingSet}(x)

\S4method{flowData}{GatingSet}(x) <- value
}
\arguments{
\item{x}{A \code{GatingSet}}

\item{value}{The replacement \code{flowSet} or \code{ncdfFlowSet} object}
}
\value{
the object with the new flowSet in place.
}
\description{
Accessor method that gets or replaces the flowset/ncdfFlowSet object in a GatingSet or GatingHierarchy
}
\details{
Accessor method that sets or replaces the ncdfFlowSet object in the GatingSet or GatingHierarchy.
}

