% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{flowJo.fasinh}
\alias{flowJo.fasinh}
\alias{flowJo.fsinh}
\title{inverse hyperbolic sine transform function}
\usage{
flowJo.fasinh(m = 4, t = 12000, a = 0.7, length = 256)

flowJo.fsinh(m = 4, t = 12000, a = 0.7, length = 256)
}
\arguments{
\item{m}{numeric the full width of the transformed display in asymptotic decades}

\item{t}{numeric the maximum value of input data}

\item{a}{numeric Additional negative range to be included in the display in asymptotic decades}

\item{length}{numeric the maximum value of transformed data}
}
\value{
fasinh/fsinh transform function
}
\description{
hyperbolic sine/inverse hyperbolic sine (flowJo-version) transform function constructor
}
\examples{
trans <- flowJo.fasinh()
data.raw <- c(1,1e2,1e3)
data.trans <- trans(data.raw)
data.trans

inverse.trans <- flowJo.fsinh()
inverse.trans(data.trans)

}
