% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{flowJo.flog}
\alias{flowJo.flog}
\title{flog transform function}
\usage{
flowJo.flog(decade = 4.5, offset = 1, max_val = 262144, min_val = 0,
  inverse = FALSE)
}
\arguments{
\item{decade}{number of decades}

\item{offset}{offset to the orignal input}

\item{max_val}{top of scale value}

\item{min_val}{lower bound of scaled value (where negative raw value gets truncated at)}

\item{inverse}{whether return the inverse function}
}
\value{
flog(or its inverse) transform function
}
\description{
flog transform function constructor. It is different from flowCore version of \link{logtGml2}
in the way that it reset negative input so that no NAN will be returned.
}
\examples{
trans <- flowJo.flog()
data.raw <- c(1,1e2,1e3)
data.trans <- trans(data.raw)
data.trans

inverse.trans <- flowJo.flog(inverse = T)
inverse.trans(data.trans)

#negative input
data.raw <- c(-10,1e2,1e3)
data.trans <- trans(data.raw)
data.trans
inverse.trans(data.trans)#we lose the original value at lower end since flog can't restore negative value

#different
trans <- flowJo.flog(decade = 3, max_val = 1e3)
data.trans <- trans(data.raw)
data.trans
inverse.trans <- flowJo.flog(decade = 3, max_val = 1e3, inverse = T)
inverse.trans(data.trans)

}
