% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{flowJo_biexp_breaks}
\alias{flowJo_biexp_breaks}
\title{flowJo biexponential breaks (integer breaks on biexp-transformed scales)}
\usage{
flowJo_biexp_breaks(n = 6, equal.space = FALSE, ...)
}
\arguments{
\item{n}{desired number of breaks (the actual number will be different depending on the data range)}

\item{equal.space}{whether breaks at equal-spaced intervals}

\item{...}{parameters passed to \code{\link[flowWorkspace]{flowJoTrans}}}
}
\value{
a function generates biexponential space
}
\description{
Used to construct \code{\link{flowJo_biexp_trans}} object
}
\examples{
data(GvHD)
fr <- GvHD[[1]]
data.raw <- exprs(fr)[, "FL1-H"]
brks.func <- flowJo_biexp_breaks(equal.space = TRUE)
brks <- brks.func(data.raw)
round(brks) # biexp space displayed at raw data scale

#transform it to verify it is equal-spaced at transformed scale
trans.func <- flowJoTrans()
brks.trans <- trans.func(brks)
round(brks.trans )
}

