% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{flowJo_fasinh_breaks}
\alias{flowJo_fasinh_breaks}
\title{flowJo inverse hyperbolic sine breaks (integer breaks on fasinh-transformed scales)}
\usage{
flowJo_fasinh_breaks(n = 6, equal.space = FALSE, ...)
}
\arguments{
\item{n}{desired number of breaks (the actual number will be different depending on the data range)}

\item{equal.space}{whether breaks at equal-spaced intervals}

\item{...}{parameters passed to flowJo.fasinh}
}
\value{
a function generates fasinh or fsinh space
}
\description{
Used to construct \code{\link{flowJo_fasinh_trans}} object
}
\examples{
data <- 1:1e3
brks.func <- flowJo_fasinh_breaks(equal.space = TRUE)
brks <- brks.func(data)
brks # fasinh space displayed at raw data scale

#transform it to verify it is equal-spaced at transformed scale
trans.func <- flowJo.fasinh()
round(trans.func(brks))
}

