% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{flowjo_flog}
\alias{flowjo_flog}
\alias{flowJo.flog}
\title{flog transform function}
\usage{
flowjo_flog(decade = 4.5, offset = 1, max_val = 262144,
  min_val = 0, scale = 1, inverse = FALSE)

flowJo.flog(decade = 4.5, offset = 1, max_val = 262144,
  min_val = 0, scale = 1, inverse = FALSE)
}
\arguments{
\item{decade}{number of decades}

\item{offset}{offset to the orignal input}

\item{max_val}{top of scale value}

\item{min_val}{lower bound of scaled value (where negative raw value gets truncated at)}

\item{scale}{the linear scale factor}

\item{inverse}{whether return the inverse function}
}
\value{
flog(or its inverse) transform function
}
\description{
flog transform function constructor. It is different from flowCore version of \link{logtGml2}
in the way that it reset negative input so that no NAN will be returned.
}
\examples{
trans <- flowjo_flog()
data.raw <- c(1,1e2,1e3)
data.trans <- trans(data.raw)
data.trans

inverse.trans <- flowjo_flog(inverse = TRUE)
inverse.trans(data.trans)

#negative input
data.raw <- c(-10,1e2,1e3)
data.trans <- trans(data.raw)
data.trans
inverse.trans(data.trans)#we lose the original value at lower end since flog can't restore negative value

#different
trans <- flowjo_flog(decade = 3, max_val = 1e3)
data.trans <- trans(data.raw)
data.trans
inverse.trans <- flowjo_flog(decade = 3, max_val = 1e3, inverse = TRUE)
inverse.trans(data.trans)

}
