% Generated by roxygen2 (4.0.1): do not edit by hand
\name{flowJoTrans}
\alias{flowJoTrans}
\title{return a biexponentioal transformation function that is compatible with flowJo}
\usage{
flowJoTrans(channelRange = 4096, maxValue = 262144, pos = 4.5, neg = 0,
  widthBasis = -10)
}
\arguments{
\item{channelRange}{\code{numeric} the maximum value of transformed data}

\item{maxValue}{\code{numeric} the maximum value of input data}

\item{pos}{\code{numeric} the full width of the transformed display in asymptotic decades}

\item{neg}{\code{numeric} Additional negative range to be included in the display in asymptotic decades}

\item{widthBasis}{\code{numeric} unkown.}
}
\description{
return a biexponentioal transformation function that is compatible with flowJo
}

