\name{flowWorkspace-internal}
\title{Internal flowWorkspace objects}
\alias{.S3MethodsClasses}
\alias{.bGate}
\alias{.booleanGate}
\alias{.children}
\alias{.ellipseFit}
\alias{.extractGate}
\alias{.flowJoTransform}
\alias{.getCalibrationTable}
\alias{.getCalibrationTableNames}
\alias{.getCalibrationTableSearch}
\alias{.getCompensationMatrices}
\alias{.getGains}
\alias{.getKeywords}
\alias{.getPopulations}
\alias{.getSampleGroups}
\alias{.getSamples}
\alias{.getStainChannels}
\alias{.hasChildren}
\alias{.hasGate}
\alias{.hasNN}
\alias{.isLeaf}
\alias{.isPopulation}
\alias{.isRoot}
\alias{.mergetables}
\alias{.parent}
\description{Internal flowWorkspace objects.}
\details{These are not to be called by the user.}
\keyword{internal}