\name{flowWorkspace-package}
\alias{flowWorkspace-package}
\alias{flowWorkspace}
\docType{package}
\title{
Import and replicate flowJo workspaces and gating schemes using flowCore.
}
\description{
Import flowJo workspaces into R. Generate the flowJo gating hierarchy and gates using flowCore functionality. Transform and compensate data in accordance with flowJo settings. Plot gates, gating hierarchies,  population statistics, and compare flowJo vs flowCore population summaries.
}
\details{
\tabular{ll}{
Package: \tab flowWorkspace\cr
Type: \tab Package\cr
Version: \tab 0.5.40\cr
Date: \tab 2011-03-04\cr
License: \tab Artistic 2.0 \cr
LazyLoad: \tab yes\cr
Depends: \tab methods, RBGL, graph, XML, flowCore, flowViz, Rgraphviz, Biobase\cr
}

}
\author{
 Greg Finak, Mike Jiang, Mose Andre

Maintainer: Greg Finak \email{gfinak@fhcrc.org}
}
\references{
\url{http://www.rglab.org/}
}
\keyword{ package }
\keyword{ utilities }
