% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoFrame.R
\name{fr_get_h5_file_path}
\alias{fr_get_h5_file_path}
\title{return the file path of underlying h5 file
For the in-memory version of cytoFrame, it returns empty string.Thus can be used to check whether it is on-disk format.}
\usage{
fr_get_h5_file_path(fr)
}
\arguments{
\item{fr}{cytoFrame object}
}
\description{
return the file path of underlying h5 file
For the in-memory version of cytoFrame, it returns empty string.Thus can be used to check whether it is on-disk format.
}
