\name{GatingSetList}
\alias{GatingSetList}
\title{Constructor for a GatingSetList}
\usage{
  GatingSetList(x, samples = NULL)
}
\arguments{
  \item{x}{a \code{list} of \code{GatingSet}s}

  \item{samples}{\code{{character} vector specifying the
  sample names if NULL, the sample names are extracted from
  GatingSets}
}
\description{
  construct a GatingSetList from a list of GatingSet
}

