% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_leaf_nodes}
\alias{get_leaf_nodes}
\title{get all the leaf nodes}
\usage{
get_leaf_nodes(x, ...)
}
\arguments{
\item{x}{GatingHierarchy/GatingSet object}

\item{...}{arguments passed to 'getNodes" method}
}
\value{
the leaf nodes
}
\description{
get all the leaf nodes
}
