% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphGatingML-methods.R
\docType{methods}
\name{getCompensationMatrices,graphGML-method}
\alias{getCompensationMatrices,graphGML-method}
\title{Extract compensation from graphGML object.}
\usage{
\S4method{getCompensationMatrices}{graphGML}(x)
}
\arguments{
\item{x}{graphGML}
}
\value{
compensation object or "FCS" when compensation comes from FCS keywords
}
\description{
Extract compensation from graphGML object.
}

