% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R
\name{getCompensationObj}
\alias{getCompensationObj}
\title{extract compensation object from GatingSet}
\usage{
getCompensationObj(gs, sampleName)
}
\arguments{
\item{gs}{GatingSet}

\item{sampleName}{sample name}
}
\description{
extract compensation object from GatingSet
}
