% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R
\name{getFullNodePath}
\alias{getFullNodePath}
\title{convert the partial gating path to the full path}
\usage{
getFullNodePath(gh, path)
}
\arguments{
\item{gh}{GatingHierarchy object}

\item{path}{the partial gating path}
}
\value{
the full gating path
}
\description{
convert the partial gating path to the full path
}
