% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphGML-methods.R
\docType{methods}
\name{getGate,graphGML,character-method}
\alias{getGate,graphGML,character-method}
\title{get gate from the node}
\usage{
\S4method{getGate}{graphGML,character}(obj, y)
}
\arguments{
\item{obj}{\code{graphGML}}

\item{y}{\code{character} node path}
}
\description{
get gate from the node
}

