\name{getIndiceMat}
\alias{getIndiceMat}
\title{Return the single-cell matrix of 1/0 dichotomized expression}
\usage{
  getIndiceMat(gh, y)
}
\arguments{
  \item{gh}{\code{GatingHierarchy} object}

  \item{y}{\code{character} node name}
}
\description{
  Return the single-cell matrix of 1/0 dichotomized
  expression
}

