% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getIndices,GatingSet,name-method}
\alias{getIndices,GatingSet,name-method}
\title{routine to return the indices by specify boolean combination of reference nodes:}
\usage{
\S4method{getIndices}{GatingSet,name}(obj, y)
}
\arguments{
\item{y}{a quoted expression.}

\item{obj}{\code{ncdfFlowSet} object}
}
\description{
It adds the boolean gates and does the gating on the fly, and
return the indices associated with that bool gate, and remove the bool gate
the typical use case would be extracting any-cytokine-expressed cells
}
\examples{
\dontrun{

	getIndices(gs,quote(`4+/TNFa+|4+/IL2+`))

}
}

