% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getMergedStats}
\alias{getMergedStats}
\title{Get Cell Population Statistics and Sample Metadata}
\usage{
getMergedStats(object, ...)
}
\arguments{
\item{object}{a \code{GatingSet} or \code{GatingSetList}}

\item{...}{additional arguments passed to \code{getPopStats}}
}
\value{
a \code{data.table} of merged population statistics with sample metadata.
}
\description{
Get Cell Population Statistics and Sample Metadata
}
\examples{
 \dontrun{
    #G is a GatingSetList
    stats = getMergedStats(G)
  }
}
