% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphGatingML-methods.R
\docType{methods}
\name{getParent,graphGML,character-method}
\alias{getParent,graphGML,character-method}
\title{get parent nodes}
\usage{
\S4method{getParent}{graphGML,character}(obj, y)
}
\arguments{
\item{obj}{\code{graphGML}}

\item{y}{\code{character} child node path}
}
\description{
get parent nodes
}

