% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSetList_Methods.R, R/GatingSet_Methods.R
\docType{methods}
\name{getSingleCellExpression,GatingSetList,character-method}
\alias{getSingleCellExpression}
\alias{getSingleCellExpression,GatingSet,character-method}
\alias{getSingleCellExpression,GatingSetList,character-method}
\title{Return the cell events data that express in any of the single populations defined in \code{y}}
\usage{
\S4method{getSingleCellExpression}{GatingSetList,character}(x, nodes, ...)

\S4method{getSingleCellExpression}{GatingSet,character}(x, nodes,
  other.markers = NULL, swap = FALSE, threshold = TRUE,
  mc.cores = getOption("mc.cores", 1L), ...)
}
\arguments{
\item{x}{A \code{GatingSet} or \code{GatingSetList} object .}

\item{nodes}{\code{character} vector specifying different cell populations}

\item{...}{other arguments

     map a named list providing the mapping between node names (as specified in the gating hierarchy of the gating set) and channel
                        names (as specified in either the \code{desc} or \code{name}
                         columns of the parameters of the associated \code{flowFrame}s
                         in the \code{GatingSet}). see examples.

     ignore.case whether to ignore case when match the marker names. Default is FALSE.}

\item{other.markers}{\code{character} vector specifying the extra markers/channels to be returned besiedes the ones derived from "nodes" and "map" argument.It is only valid when threshold is set to FALSE.}

\item{swap}{\code{logical} indicates whether channels and markers of flow data are swapped.}

\item{threshold}{\code{logical} indicates whether to threshold the flow data by setting intensity value to zero when it is below the gate threshold.}

\item{mc.cores}{passed to \code{mclapply}. Default is 1, which means the process runs in serial mode. When it is larger than 1, parallel mode is enabled.}
}
\value{
A \code{list} of \code{numerci matrices}
}
\description{
Returns a list of matrix containing the events that expressed in any one of the populations efined in \code{y}
}
\examples{
\dontrun{
  #G is a GatingSet
	geData(G,3)
	res <- getSingleCellExpression(gs[1], c("4+/TNFa+", "4+/IL2+"))
	res[[1]]
 # if it fails to match the given nodes to the markers, then try to provide the mapping between node and marker explicitly
	res <- getSingleCellExpression(gs[1], c("4+/TNFa+", "4+/IL2+") , map = list("4+/TNFa+" = "TNFa", "4+/IL2+" = "IL2"))
}
}
\author{
Mike Jiang \email{wjiang2@fhcrc.org}
}
\seealso{
\code{\link{getIndices}}  \code{\link{getPopStats}}
}

