% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\alias{getStats.GatingSetList}
\alias{getStats.GatingSet}
\alias{getStats.GatingHierarchy}
\title{Exact MFI from populations(or nodes) for all the markers}
\usage{
getStats(x, ...)

\method{getStats}{GatingSetList}(x, ...)

\method{getStats}{GatingSet}(x, ...)

\method{getStats}{GatingHierarchy}(x, nodes = NULL, type = "count",
  inverse.transform = FALSE, stats.fun.arg = list(), ...)
}
\arguments{
\item{x}{a GatingSet or GatingHierarchy}

\item{...}{arguments passed to \link{getNodes} method.}

\item{nodes}{the character vector specifies the populations of interest. default is all available nodes}

\item{type}{the character vector specifies the type of pop stats or
a function used to compute population stats.
when character, it is expected to be either "count" or "percent". Default is "count" (total number of events in the populations).
when a function,  it takes a flowFrame object through 'fr' argument and return the stats as a named vector.}

\item{inverse.transform}{logical flag . Whether inverse transform the data before computing the stats.}

\item{stats.fun.arg}{a list of arguments passed to `type` when 'type' is a function.}
}
\value{
a data.table that contains MFI values for each marker per column along with 'pop' column and 'sample' column (when used on a 'GatingSet')
}
\description{
It calculates the MFI for each marker.
}
