% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphGML-methods.R
\docType{methods}
\name{getTransformations,graphGML-method}
\alias{getTransformations,graphGML-method}
\title{Extract transformations from graphGML object.}
\usage{
\S4method{getTransformations}{graphGML}(x)
}
\arguments{
\item{x}{graphGML}
}
\value{
transformList object
}
\description{
Extract transformations from graphGML object.
}

