% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{gh_apply_to_new_fcs}
\alias{gh_apply_to_new_fcs}
\alias{GatingSet,GatingHierarchy,character-method}
\title{constructors for GatingSet}
\usage{
gh_apply_to_new_fcs(x, files, swap_cols = FALSE, ...)
}
\arguments{
\item{x}{GatingHierarchy}

\item{files}{fcs file paths}

\item{swap_cols}{for internal usage}

\item{...}{other arguments.}

\item{y}{sample names}

\item{path}{\code{character} specifies the path to the flow data (FCS files)}
}
\description{
construct object from existing gating hierarchy(gating template) and flow data
}
