% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R
\name{gh_get_cluster_labels}
\alias{gh_get_cluster_labels}
\title{Retrieve the cluster labels from the cluster nodes}
\usage{
gh_get_cluster_labels(gh, parent, cluster_method_name)
}
\arguments{
\item{gh}{GatingHierarchy}

\item{parent}{the parent population/node name or path}

\item{cluster_method_name}{the name of the clustering method}
}
\description{
Clustering results are stored as individual gated nodes. This helper function
collect all the gating indices from the same clustering run (identified by 'parent'
node and 'cluster_method_name" and merge them as a single factor.
}
