% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSingleCellExpression.R
\name{gh_get_indices_mat}
\alias{gh_get_indices_mat}
\title{Return the single-cell matrix of 1/0 dichotomized expression}
\usage{
gh_get_indices_mat(gh, y)
}
\arguments{
\item{gh}{\code{GatingHierarchy} object}

\item{y}{\code{character} string containing the boolean or of node names.e.g. 'cd4|cd8'}
}
\description{
Return the single-cell matrix of 1/0 dichotomized expression
}
