% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getStats}
\alias{getStats}
\title{Exact MFI from populations(or nodes) for all the markers}
\usage{
getStats(x, ...)
}
\arguments{
\item{x}{a GatingSet or GatingHierarchy}

\item{...}{arguments passed to \link{gs_get_pop_paths} method.}
}
\value{
a data.table that contains MFI values for each marker per column along with 'pop' column and 'sample' column (when used on a 'GatingSet')
}
\description{
It calculates the MFI for each marker.
}
\examples{
\dontrun{
dataDir <- system.file("extdata",package="flowWorkspaceData")
suppressMessages(gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE)))

# get stats all nodes
dt <- gs_get_stats(gs) #default is "count"

nodes <- c("CD4", "CD8")
gs_get_stats(gs, nodes, "percent")

# pass a build-in function
gs_get_stats(gs, nodes, type = pop.MFI)

# compute the stats based on the raw data scale
gs_get_stats(gs, nodes, type = pop.MFI, inverse.transform = TRUE)

# supply user-defined stats fun
pop.quantiles <- function(fr){
   chnls <- colnames(fr)
   res <- matrixStats::colQuantiles(exprs(fr), probs = 0.75)
   names(res) <- chnls
   res
   }
gs_get_stats(gs, nodes, type = pop.quantiles)
}
}
