% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R
\name{gh_pop_get_cluster_name}
\alias{gh_pop_get_cluster_name}
\title{check if a node is clustering node}
\usage{
gh_pop_get_cluster_name(gh, node)
}
\arguments{
\item{gh}{GatingHierarchy}

\item{node}{the population/node name or path}
}
\value{
the name of the clustering method. If it is not cluster node, returns NULL
}
\description{
check if a node is clustering node
}
