% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R
\name{gh_pop_get_full_path}
\alias{gh_pop_get_full_path}
\title{convert the partial gating path to the full path}
\usage{
gh_pop_get_full_path(gh, path)
}
\arguments{
\item{gh}{GatingHierarchy object}

\item{path}{the partial gating path}
}
\value{
the full gating path
}
\description{
convert the partial gating path to the full path
}
