% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSingleCellExpression.R
\name{gh_pop_get_indices_mat}
\alias{gh_pop_get_indices_mat}
\title{Return the single-cell matrix of 1/0 dichotomized expression}
\usage{
gh_pop_get_indices_mat(gh, y)
}
\arguments{
\item{gh}{\code{GatingHierarchy} object}

\item{y}{\code{character} vector containing the node names}
}
\description{
Return the single-cell matrix of 1/0 dichotomized expression
}
