% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{getProp}
\alias{getProp}
\alias{gh_pop_get_proportion}
\alias{getTotal}
\alias{gh_pop_get_count}
\title{Get count or proportion from populations}
\usage{
getProp(x, y, xml = FALSE)

gh_pop_get_proportion(x, y, xml = FALSE)

getTotal(x, y, xml = FALSE)

gh_pop_get_count(x, y, xml = FALSE)
}
\arguments{
\item{x}{GatingHierarchy}

\item{y}{\code{character} node name or path}

\item{xml}{whether to extract xml stats or openCyto stats}
}
\description{
Get count or proportion from populations
}
