% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{clone}
\alias{clone}
\alias{clone-methods}
\alias{clone,GatingSet-method}
\alias{gs_clone}
\alias{gs_copy_tree_only}
\title{clone a GatingSet}
\usage{
clone(x, ...)

gs_clone(x, h5_dir = tempdir())

gs_copy_tree_only(x)
}
\arguments{
\item{x}{A \code{GatingSet}}

\item{...}{h5_dir = tempdir() the directory to store the h5-based flow data matrix}

\item{h5_dir}{h5 dir for the new gs}
}
\value{
A copy of a given \code{GatingSet}.
}
\description{
clone a GatingSet
}
\details{
Note that the regular R assignment operation on a \code{GatingSet} object does not return the copy as
  one would normally expect because the \code{GatingSet} contains environment slots (and external pointer for \code{GatingSet}),
  which require deep-copying. So make sure to use this clone method in order to make a copy of existing object.
}
\examples{
  \dontrun{
    #gs is  a GatingSet
    gs2 <-gs_clone(gs) #gs2 is independent from gs and have its own copy of both gating trees and flow data
    gs3 <- gs_copy_tree_only(gs) #gs3 has its own copy of gating trees but share the same flow data with original gs
  }
}
