% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{gs_compute_stats}
\alias{gs_compute_stats}
\title{Compute all the stats}
\usage{
gs_compute_stats(gs)
}
\arguments{
\item{gs}{GatingSet}
}
\description{
To save time, some channel-specific stats such as mean, SD are not computed automatically at the time when they are added to GatingSet.
This function allows usesr to invoke this computation manually if they want to extract these values.
}
