% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{gs_cyto_data}
\alias{gs_cyto_data}
\alias{flowData}
\alias{flowData<-}
\alias{gs_cyto_data<-}
\alias{flowData,GatingSet-method}
\alias{flowData<-,GatingSet-method}
\alias{gs_cyto_data,GatingSet-method}
\alias{gs_cyto_data<-,GatingSet-method}
\title{Fetch or replace the flowData object associated with a GatingSet .}
\usage{
gs_cyto_data(x, ...)

\S4method{gs_cyto_data}{GatingSet}(x, inverse.transform = FALSE)

gs_cyto_data(x) <- value
}
\arguments{
\item{x}{A \code{GatingSet}}

\item{...}{other arugments}

\item{inverse.transform}{logical flag indicating whether to inverse transform the data}

\item{value}{The replacement \code{flowSet} or \code{ncdfFlowSet} object}
}
\value{
the object with the new flowSet in place.
}
\description{
Accessor method that gets or replaces the \code{\link{cytoset}}/\code{\link[flowCore]{flowSet}}/\code{\link[ncdfFlow:ncdfFlowSet-class]{ncdfFlowSet}} object in a GatingSet or GatingHierarchy
}
\details{
Accessor method that sets or replaces the ncdfFlowSet object in the GatingSet or GatingHierarchy.
}
