% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R
\name{gs_get_compensation_internal}
\alias{gs_get_compensation_internal}
\title{extract compensation object from GatingSet}
\usage{
gs_get_compensation_internal(gs, sampleName)
}
\arguments{
\item{gs}{GatingSet}

\item{sampleName}{sample name}
}
\description{
extract compensation object from GatingSet
}
