% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{gs_get_phylo}
\alias{gs_get_phylo}
\alias{gh_get_phylo}
\title{Get a ape::phylo object from a \code{\link{GatingSet}} or \code{\link{GatingHierarchy}}}
\usage{
gs_get_phylo(gs, below = "root")

gh_get_phylo(gs, below = "root")
}
\description{
Get a ape::phylo object from a \code{\link{GatingSet}} or \code{\link{GatingHierarchy}}
}
