% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\docType{methods}
\name{getPopStats,GatingSet-method}
\alias{getPopStats,GatingSet-method}
\alias{gs_get_pop_stats}
\title{Return a table of population statistics for all populations in a GatingHierarchy/GatingSet
  or the population proportions or the total number of events of a node (population) in a GatingHierarchy}
\usage{
\S4method{getPopStats}{GatingSet}(x, statistic = c("freq", "count"),
  xml = FALSE, subpopulations = NULL, format = c("long", "wide"),
  path = "full", ...)

gs_get_pop_stats(x, statistic = c("freq", "count"), xml = FALSE,
  subpopulations = NULL, format = c("long", "wide"), path = "full",
  ...)
}
\arguments{
\item{x}{A \code{GatingHierarchy} or \code{GatingSet}}

\item{statistic}{\code{character} specifies the type of population statistics to extract.(only valid when format is "wide"). Either "freq" or "count" is currently supported.}

\item{xml}{\code{logical} indicating whether the statistics come from xml (if parsed from xml workspace) or from openCyto.}

\item{subpopulations}{\code{character} vector to specify a subset of populations to return. (only valid when format is "long")}

\item{format}{\code{character} value of c("wide", "long") specifing whether to origanize the output in long or wide format}

\item{path}{\code{character} see \link{gs_get_pop_paths}}

\item{...}{Additional arguments passed to \link{gs_get_pop_paths}}
}
\value{
gs_get_pop_stats returns a \code{data.frame} with columns for the population name, xml derived counts, openCyto derived counts, and the population proportions (relative to their parent pouplation).
gh_get_proportion returns  a population frequency \code{numeric}.
gh_get_count returns a \code{numeric} value of the total number of elements in the population.
}
\description{
gh_get_proportion calculates the population proportion (events in the gate / events in the parent population) associated with a node in the \code{GatingHierarchy}.
gs_get_pop_stats is more useful than getPop. Returns a table of population statistics for all populations in a \code{GatingHierarchy}/\code{GatingSet}. Includes the xml counts, openCyto counts and frequencies.
gh_get_count returns the total number of events in the gate defined in the GatingHierarchy object
}
\details{
gs_get_pop_stats returns a table population statistics for all populations in the gating hierarchy. The output is useful for verifying that the import was successful, if the xml and openCyto derived counts don't differ much (i.e. if they have a small coefficient of variation.) for a GatingSet, returns a matrix of proportions for all populations and all samples
gh_get_proportion returns the proportion of cells in the gate, relative to its parent.
gh_get_count returns the total number of events included in this gate. The contents of "thisTot" variable in the "metadata" environment of the \code{nodeData} element associated with the gating tree and gate / population.
}
\examples{
        \dontrun{
        #gh is a GatingHierarchy
        gs_get_pop_stats(gh);
        #proportion for the fifth population
        gh_get_proportion(gh,gs_get_pop_paths(gh)[5])
        gh_get_count(gh,gs_get_pop_paths(gh,tsort=T)[5])

        #gs is a GatingSet
        gs_get_pop_stats(gs)
        #optionally output in long format as a data.table
        gs_get_pop_stats(gs, format = "long", path = "auto")
        #only get stats for a subset of populations
        gs_get_pop_stats(gs, format = "long", subpopulations = gs_get_pop_paths(gs)[4:6])
        }
}
\seealso{
\code{\link{gs_get_pop_paths}}
}
