% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{gs_is_persistent}
\alias{gs_is_persistent}
\alias{gs_is_h5}
\alias{isNcdf}
\title{determine whether the flow data associated with a GatingSet is persistent(on-disk) or in-memory}
\usage{
gs_is_persistent(x)

gs_is_h5(x)

isNcdf(x)
}
\arguments{
\item{x}{\code{GatingSet} object}
}
\value{
\code{logical}
}
\description{
determine whether the flow data associated with a GatingSet is persistent(on-disk) or in-memory
}
