% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{gs_pop_get_gs}
\alias{gs_pop_get_gs}
\title{subset gs by population node}
\usage{
gs_pop_get_gs(gs, pop)
}
\arguments{
\item{gs}{GatingSet}

\item{pop}{the population node that will become the new root node}
}
\value{
a new GatingSet that share the underlying events data
}
\description{
Basically it returns a new GatingSet with only the substree of the given population node
}
