% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{gs_pop_stats_compute}
\alias{gs_pop_stats_compute}
\title{Compute all the stats}
\usage{
gs_pop_stats_compute(gs, node = "root")
}
\arguments{
\item{gs}{GatingSet}

\item{node}{the population node. Default is root.}
}
\description{
To save time, some channel-specific stats such as mean, SD are not computed automatically at the time when they are added to GatingSet.
This function allows users to invoke this computation manually for the given population node and all its descendants.
}
