% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_GatingSet.R
\name{dropRedundantChannels}
\alias{dropRedundantChannels}
\alias{gs_remove_redundant_channels}
\title{Remove the channels from flow data that are not used by gates}
\usage{
dropRedundantChannels(...)

gs_remove_redundant_channels(gs, ...)
}
\arguments{
\item{...}{other arugments passed to gs_get_pop_paths method}

\item{gs}{a \code{GatingSet}}
}
\value{
a new \code{GatingSet} object that has redundant channels removed. Please note that this new object shares the same reference (or external pointers) with the original GatingSets.
}
\description{
Removing these redundant channels can help standardize the channels across different GatingSet objects and make them mergable.
}
\examples{
\dontrun{
gs_new <- gs_remove_redundant_channels(gs)
}
}
