% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R,
%   R/GatingSet_Methods.R
\docType{methods}
\name{setNode,GatingHierarchy,character,logical-method}
\alias{setNode,GatingHierarchy,character,logical-method}
\alias{gh_set_node_visible}
\alias{gs_set_node_visible}
\title{hide/unhide a node}
\usage{
\S4method{setNode}{GatingHierarchy,character,logical}(x, y, value)

gh_set_node_visible(x, y, value)

gs_set_node_visible(x, y, value)
}
\arguments{
\item{x}{\code{GatingHierarchy} object}

\item{y}{\code{character} node name or path}
}
\description{
hide/unhide a node
}
\examples{
\dontrun{
     gh_set_node_visible(gh, 4, FALSE) # hide a node
     gh_set_node_visible(gh, 4, TRUE) # unhide a node
}
}
