% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_GatingSet.R
\name{gs_split_by_tree}
\alias{gs_split_by_tree}
\alias{groupByTree}
\title{split GatingSets into groups based on their gating schemes
Be careful that the splitted resluts still points to the original data set!!}
\usage{
gs_split_by_tree(x)
}
\arguments{
\item{x}{a list of GatingSets or one GatingSet}
}
\value{
when x is a GatingSet, this function returns a list of sub-GatingSets
When x is a list of GatingSets, it returns a list of list, each list itself is a list of GatingSets, which share the same gating tree.
}
\description{
It allows isomorphism in Gating tree and ignore difference in hidden nodes
i.e. tree is considered to be the same
as long as gs_get_pop_paths(gh, path = "auto", showHidden = F) returns the same set
}
\examples{
\dontrun{
gslist <- list(gs1, gs2, gs3, gs4, gs5)
gs_groups <- gs_split_by_tree(gslist)
}
}
