% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{gscluster}
\alias{gscluster}
\title{construct a gscluster object from an existing gatingset object}
\usage{
gscluster(gs, path = tempfile())
}
\arguments{
\item{gs}{GatingSet}

\item{path}{the path to store the on-disk portion of gs (i.e. pb and cytoframe content)}
}
\description{
construct a gscluster object from an existing gatingset object
}
