% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSetList_Methods.R
\docType{methods}
\name{rbind2,GatingSetList,missing-method}
\alias{rbind2,GatingSetList,missing-method}
\alias{gslist_to_gs}
\title{Merge a GatingSetList into a single GatingSet}
\usage{
\S4method{rbind2}{GatingSetList,missing}(x, y = "missing", ...)

gslist_to_gs(x, ...)
}
\arguments{
\item{x}{GatingSetList}

\item{y}{\code{missing} not used.}

\item{...}{other arguments passed to \code{gslist_to_gs} method for \code{ncdfFlowList}}
}
\description{
Merge a GatingSetList into a single GatingSet
}
