% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\docType{methods}
\name{identifier-methods}
\alias{identifier-methods}
\alias{identifier}
\alias{identifier,GatingSet-method}
\alias{identifier,GatingSetList-method}
\alias{identifier<-,GatingSet,ANY-method}
\alias{identifier<-,GatingSet-method}
\alias{identifier<-,GatingSet,character-method}
\title{Retrieve/replace the GUID of a GatingSet or GatingSetList}
\usage{
identifier(object)

identifier(object) <- value
}
\arguments{
\item{object}{a GatingSet or GatingSetList}

\item{value}{string}
}
\description{
Retrieve or replace the GUID (globally unique identifier) for a GatingSet or GatingSetList
}
