% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet-Methods.R
\name{insertGate}
\alias{insertGate}
\title{insert a dummy gate to the GatingSet}
\usage{
insertGate(gs, gate, parent, children)
}
\arguments{
\item{gs}{\code{GatingSet} to work with}

\item{gate}{\code{filter} a dummy gate to be inserted, its 'filterId' will be used as the population name}

\item{parent}{\code{character} full path of parent node where the new dummy gate to be added to}

\item{children}{\code{character} full path of chidlren nodes that the new dummy gate to be parent of}
}
\value{
a new \code{GatingSet} object with the new gate added but share the same flow data with the input 'GatingSet'
}
\description{
Is is useful trick to make the tree structure of GatingSet same with other so that
they can be combined into a 'GatingSetList' object.
}
\examples{
\dontrun{
#construct a dummy singlet gate
 dummyGate <- rectangleGate("FSC-A" = c(-Inf, Inf), "FSC-H" = c(-Inf, Inf), filterId = "singlets")
#insert it between the 'not debris" node and "lymph" node
 gs_clone <- insertGate(gs, dummyGate, "not debris", "lymph")
}
}

