% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R, R/cytoset.R
\name{cf_is_subsetted}
\alias{cf_is_subsetted}
\alias{cs_is_subsetted}
\title{check whether a cytoframe/cytoset is a subsetted(by column or by row) view}
\usage{
cf_is_subsetted(x)

cs_is_subsetted(x)
}
\arguments{
\item{x}{a cytoset or cytoframe}
}
\description{
check whether a cytoframe/cytoset is a subsetted(by column or by row) view
}
