% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{is_tiledb_support}
\alias{is_tiledb_support}
\title{check whether cytolib is build with tiledb support}
\usage{
is_tiledb_support()
}
\value{
TRUE or FALSE
}
\description{
check whether cytolib is build with tiledb support
}
