\name{isNcdf}
\alias{isNcdf}
\title{determine the flow data associated with a Gating Hiearchy is based on `ncdfFlowSet` or `flowSet`}
\usage{
  isNcdf(x)
}
\arguments{
  \item{x}{\code{GatingHiearchy} object}
}
\value{
  \code{logical}
}
\description{
  determine the flow data associated with a Gating Hiearchy
  is based on `ncdfFlowSet` or `flowSet`
}

