% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{lapply,GatingSet-method}
\alias{lapply,GatingSet-method}
\title{apply \code{FUN} to each sample (i.e. \code{GatingHierarchy})}
\usage{
\S4method{lapply}{GatingSet}(X, FUN, ...)
}
\arguments{
\item{X}{
    A list-like or vector-like object.
  }

\item{FUN}{
    See \code{?base::\link[base]{lapply}} for a description of
    these arguments.
  }
}
\description{
sample names are used for names of the returned list
}

