% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{length,GatingSet-method}
\alias{length,GatingSet-method}
\alias{length-methods}
\title{Methods to get the length of a GatingSet}
\usage{
\S4method{length}{GatingSet}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Return the length of a \code{GatingSet} or \code{GatingSetList} object (number of samples).
}

