% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoFrame.R
\name{load_cytoframe_from_h5}
\alias{load_cytoframe_from_h5}
\title{Load the cytoFrame from h5 format}
\usage{
load_cytoframe_from_h5(filename, on_disk = TRUE)
}
\arguments{
\item{filename}{the full path of the output h5 file}

\item{on_disk}{logical flag indicating whether to keep the data on disk and load it on demand. Default is TRUE.}
}
\description{
Load the cytoFrame from h5 format
}
