% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{load_gs_cluster}
\alias{load_gs_cluster}
\title{construct a gscluster object from an existing gatingset archive folder}
\usage{
load_gs_cluster(path, ...)
}
\arguments{
\item{path}{a gs archive folder.}

\item{...}{see 'load_gs()'}
}
\description{
construct a gscluster object from an existing gatingset archive folder
}
