% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R, R/cytoset.R
\name{load_meta}
\alias{load_meta}
\alias{cf_flush_meta}
\alias{flush_meta}
\alias{cf_load_meta}
\alias{cs_flush_meta}
\alias{cs_load_meta}
\title{Flush/load meta data (keywords, pData, channels/markers) to/from disk (only valid for on-disk cytoset/cytoframe)}
\usage{
cf_flush_meta(cf)

cf_load_meta(cf)

cs_flush_meta(cs)

cs_load_meta(cs)
}
\arguments{
\item{cf}{cytoframe object}

\item{cs}{cytoset object}
}
\description{
Flush/load meta data (keywords, pData, channels/markers) to/from disk (only valid for on-disk cytoset/cytoframe)
}
