% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R, R/cytoset.R
\name{lock}
\alias{lock}
\alias{cf_lock}
\alias{cf_unlock}
\alias{cs_lock}
\alias{cs_unlock}
\title{Lock/Unlock the cytoset/cytoframe by turning on/off its read-only flag}
\usage{
cf_lock(cf)

cf_unlock(cf)

cs_lock(cs)

cs_unlock(cs)
}
\arguments{
\item{cf}{cytoframe object}

\item{cs}{cytoset object}
}
\description{
Lock/Unlock the cytoset/cytoframe by turning on/off its read-only flag
}
