% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{logicle_breaks}
\alias{logicle_breaks}
\title{flowCore logicle breaks (integer breaks on logicle-transformed scales)}
\usage{
logicle_breaks(n = 6, equal.space = FALSE, ...)
}
\arguments{
\item{n}{desired number of breaks (the actual number will be different depending on the data range)}

\item{equal.space}{whether breaks at equal-spaced intervals}

\item{...}{parameters passed to flowJo.fasinh}
}
\value{
a function that generates logicle space
}
\description{
Used to construct \code{\link{logicle_trans}} object
}
\examples{
data <- 1:1e3
brks.func <- logicle_breaks(equal.space = TRUE)
brks <- brks.func(data)
brks # logicle space displayed at raw data scale

#logicle transform it to verify it is equal-spaced at transformed scale
trans.obj <- logicleTransform()
trans.func <- slot(trans.obj, ".Data")
brks.trans <- trans.func(brks)
brks.trans
}

