% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{logicle_trans}
\alias{logicle_trans}
\title{logicle transformation.}
\usage{
logicle_trans(..., n = 6, equal.space = FALSE)
}
\arguments{
\item{...}{arguments passed to logicleTransform.}

\item{n}{desired number of breaks (the actual number will be different depending on the data range)}

\item{equal.space}{whether breaks at equal-spaced intervals}
}
\value{
a logicle transformation object
}
\description{
Used for construct logicle transform object.
}
\examples{
trans.obj <- logicle_trans(equal.space = TRUE)
data <- 1:1e3
brks.func <- trans.obj[["breaks"]]
brks <- brks.func(data)
brks # logicle space displayed at raw data scale
#transform it to verify the equal-spaced breaks at transformed scale
print(trans.obj[["transform"]](brks))
}
