% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{get_log_level}
\alias{get_log_level}
\alias{set_log_level}
\title{get/set the log level}
\usage{
get_log_level()

set_log_level(level = "none")
}
\arguments{
\item{level}{a \code{character} that represents the log level
, can be value of c("none", "GatingSet", "GatingHierarchy", "Population", "gate")
   default is "none" , which does not print any information from C parser.}
}
\value{
a character that represents the internal log level
}
\description{
It is helpful sometime to get more detailed print out for the purpose of trouble shooting
}
\examples{
get_log_level()
set_log_level("Population")
get_log_level()

}
