% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flow_trans.R
\name{logtGml2_trans}
\alias{logtGml2_trans}
\title{Gating-ML 2.0 Log transformation.}
\usage{
logtGml2_trans(M = 4.5, n = 6, equal.space = FALSE)
}
\arguments{
\item{M}{number of decades}

\item{n}{desired number of breaks (the actual number will be different depending on the data range)}

\item{equal.space}{whether breaks at equal-spaced intervals}
}
\value{
logtGml2 transformation object
}
\description{
Used to construct flog transformer object.
(which uses a specilized \link{flowjo_flog})
}
\examples{
trans.obj <- logtGml2_trans(M = 1, equal.space = TRUE)
data <- 1:1e3
brks.func <- trans.obj[["breaks"]]
brks <- brks.func(data)
brks # fasinh space displayed at raw data scale

#transform it to verify it is equal-spaced at transformed scale
trans.func <- trans.obj[["transform"]]
brks.trans <- trans.func(brks)
brks.trans
}
