% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R
\docType{methods}
\name{markernames}
\alias{markernames}
\alias{markernames,GatingSet-method}
\alias{markernames,GatingHierarchy-method}
\alias{markernames<-,GatingHierarchy-method}
\alias{markernames<-,GatingSet,ANY-method}
\alias{markernames<-,GatingSet-method}
\alias{colnames,GatingHierarchy-method}
\alias{colnames,GatingSet-method}
\alias{colnames<-,GatingHierarchy-method}
\alias{colnames<-,GatingSet,ANY-method}
\alias{colnames<-,GatingSet-method}
\title{Get/set the column(channel) or marker names}
\usage{
markernames(object)

markernames(object) <- value

colnames(object)

colnames(object) <- value
}
\arguments{
\item{value}{named character vector for markernames<-, regular character vector for colnames<-.}

\item{x, object}{GatingHierarchy/GatingSet/GatingSetList}

\item{do.NULL, prefix}{not used.}
}
\description{
It simply calls the methods for the underlying flow data (flowSet/ncdfFlowSet/ncdfFlowList).
}
\examples{
\dontrun{

markers.new <- c("CD4", "CD8")
chnls <- c("<B710-A>", "<R780-A>")
names(markers.new) <- chnls
markernames(gs) <- markers.new

chnls <- colnames(gs)
chnls.new <- chnls
chnls.new[c(1,4)] <- c("fsc", "ssc")
colnames(gs) <-  chnls.new
}
}
