% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\name{matchPath}
\alias{matchPath}
\title{Given the leaf node, try to find out if a collection of nodes can be matched to a path in a graph(tree) by the bottom-up searching}
\usage{
matchPath(g, leaf, nodeSet)
}
\arguments{
\item{g}{graphNEL}

\item{leaf}{the name of leaf(terminal) node}

\item{nodeSet}{a set of node names}
}
\value{
TRUE if path is found, FALSE if not path is matched.
}
\description{
Given the leaf node, try to find out if a collection of nodes can be matched to a path in a graph(tree) by the bottom-up searching
}

