% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\name{merge.counts}
\alias{merge.counts}
\title{compare the counts to cytobank's exported csv so that the parsing result can be verified.}
\usage{
\method{merge}{counts}(gs, file, id.vars = c("FCS Filename", "population"))
}
\arguments{
\item{gs}{parsed GatingSet}

\item{file}{the stats file (contains the populatio counts) exported from cytobank.}

\item{id.vars}{either "population" or "FCS filename" that tells whether the stats file format is one population per row or FCS file per row.}
}
\value{
a data.table (in long format) that contains the counts from openCyto and Cytobank side by side.
}
\description{
compare the counts to cytobank's exported csv so that the parsing result can be verified.
}

