% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_gslist.R
\name{merge_list_to_gs}
\alias{merge_list_to_gs}
\title{Merge a list of GatingSets into a single GatingSet}
\usage{
merge_list_to_gs(x, ...)
}
\arguments{
\item{x}{a list of GatingSets}

\item{...}{other arguments (not used)}
}
\description{
It also checks the consistency of the cyto data and gates.
}
