% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveNode.R
\name{moveNode}
\alias{moveNode}
\title{move a node along with all of its descendant nodes to the given ancester}
\usage{
moveNode(gh, node, to)
}
\arguments{
\item{gh}{GatingHierarchy}

\item{node}{the node to be moved}

\item{to}{the new parent node under which the \code{node} will be moved to}
}
\description{
move a node along with all of its descendant nodes to the given ancester
}
\examples{
library(flowWorkspace)
dataDir <- system.file("extdata",package="flowWorkspaceData")
suppressMessages(gs <- load_gs(list.files(dataDir, pattern = "gs_manual",full = TRUE)))
gh <- gs[[1]]
old.parent <- getParent(gh, "CD4")
new.parent <- "singlets"
moveNode(gh, "CD4", new.parent)
getParent(gh, "CD4")
}
