% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\name{parse.gateInfo}
\alias{parse.gateInfo}
\title{Parse the cytobank custom_info for each gate}
\usage{
parse.gateInfo(file, ...)
}
\arguments{
\item{file}{xml file path}

\item{...}{additional arguments passed to the handlers of 'xmlTreeParse'}
}
\value{
a data.frame that contains three columns: id (gateId), name (gate name), fcs (fcs_file_filename).
}
\description{
Fcs filename and gate name stored in 'custom_info' element are beyong the scope of
the gatingML standard and thus not covered by the default 'read.gatingML'.
}

