% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\name{parse.gatingML}
\alias{parse.gatingML}
\title{A wrapper that parse the gatingML and FCS files into GatingSet}
\usage{
parse.gatingML(xml, FCS)
}
\arguments{
\item{xml}{the full path of gatingML file}

\item{FCS}{FCS files to be loaded}
}
\value{
a GatingSet
}
\description{
A wrapper that parse the gatingML and FCS files into GatingSet
}

