% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphGML-methods.R
\docType{methods}
\name{plot,graphGML,missing-method}
\alias{plot,graphGML,missing-method}
\title{plot the population tree stored in graphGML.}
\usage{
\S4method{plot}{graphGML,missing}(x, y = "missing", label = c("popName",
  "gateName"))
}
\arguments{
\item{x}{a graphNEL generated by constructTree function}

\item{y}{not used}

\item{label}{specifies what to be dispaled as node label. Can be either 'popName' (population name parsed from GateSets) or 'gateName'(the name of the actual gate associated with each node)}
}
\description{
The node with dotted order represents the population that has tailored gates (sample-specific gates) defined.
}
\examples{
\dontrun{
xmlfile <- system.file("extdata/cytotrol_tcell_cytobank.xml", package = "flowWorkspace")
g <- read.gatingML.cytobank(xmlfile)
plot(g)

}
}

