% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_GatingSet.R
\name{plot_diff_tree}
\alias{plot_diff_tree}
\title{visualize the tree structure differnece among the GatingSets}
\usage{
plot_diff_tree(x, path = "auto", ...)
}
\arguments{
\item{x}{\code{list} of groups(each group is a list of 'GatingSet`). it is usually the outcome from \link{groupByTree}.}

\item{path}{passed to \code{getNodes}}

\item{...}{passed to \code{getNodes}}
}
\description{
visualize the tree structure differnece among the GatingSets
}
\examples{
\dontrun{
gslist <- list(gs1, gs2, gs3, gs4, gs5)
gs_groups <- groupByTree(gslist)
plot_diff_tree(gs_groups)
}
}
