% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingHierarchy-Methods.R
\name{prettyAxis}
\alias{prettyAxis}
\title{Determine tick mark locations and labels for a given channel axis}
\usage{
prettyAxis(gh, channel)
}
\arguments{
\item{gh}{\code{GatingHiearchy}}

\item{channel}{\code{character} channel name}
}
\value{
when there is transformation function associated with the given channel, it returns a list of that contains positions and labels to draw on the axis
 other wise returns NULL
}
\description{
Determine tick mark locations and labels for a given channel axis
}
\examples{
\dontrun{
 prettyAxis(gh, "<B710-A>")
}
}

