\name{[<-,GatingSet,ANY,ANY,GatingSet-method}
\alias{[<-,GatingSet,ANY,ANY,GatingSet-method}
\title{
Replace parts of a GatingSet
}
\description{
Replace part of a GatingSet
}
\arguments{
  \item{x}{
A \code{GatingSet} object
}
  \item{i}{
The index to be replaced
}

  \item{\dots}{
Additional arguments
}
  \item{value}{
	A \code{GatingSet} object of size \code{length(i)}.
}
}
\keyword{internal}