\name{[[<-,GatingSet,ANY,ANY,GatingHierarchy-method}
\alias{[[<-,GatingSet,ANY,ANY,GatingHierarchy-method}
\title{
	Replace parts of a GatingSet
}
\description{
	Replace parts of a GatingSet
}
\arguments{
  \item{x}{
A \code{GatingSet} object
}
  \item{i}{
The index to be replaced
}
  \item{\dots}{
Additional arguments. Not used.
}
  \item{value}{
	A \code{GatingHierarchy} object
	
}
}
