% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gs.R
\name{sampleNames,character-method}
\alias{sampleNames,character-method}
\title{Get sample names from a GatingSet archive folder}
\usage{
\S4method{sampleNames}{character}(object)
}
\arguments{
\item{object}{a \code{GatingSet} folder}
}
\value{
A character vector of sample names
}
\description{
retrive  sample names by scanning h5 files from a GatingSet folder
}
\examples{
      \dontrun{
        sampleNames(gsdir)
      }
}
