% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsexperiment.R
\name{sce_to_gs}
\alias{sce_to_gs}
\title{convert SingleCellExperiment to a GatingSet}
\usage{
sce_to_gs(
  sce,
  assay_type = "intensity",
  sample = "sample",
  channel = NA,
  marker = NA,
  ...
)
}
\arguments{
\item{sce}{SingleCellExperiment object}

\item{assay_type}{select assay to use}

\item{sample}{select the column from colData to represent sample ID}

\item{channel}{select the column from rowData to represent channels}

\item{marker}{select the column from rowData to represent markers}
}
\description{
note that only selected assay and rowData columns will be extracted into the GatingSet
}
