% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set.count.xml}
\alias{set.count.xml}
\title{save the event counts parsed from xml into c++ tree structure}
\usage{
set.count.xml(gh, node, count)
}
\arguments{
\item{gh}{GatingHierarchy}

\item{node}{the unique gating path that uniquely identifies a population node}

\item{count}{integer number that is events count for the respective gating node directly parsed from xml file}
}
\description{
It is for internal use by the diva parser
}
\examples{
\dontrun{
set.count.xml(gh, "CD3", 10000)
}
}
