% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{setNode,GatingHierarchy,character,character-method}
\alias{setNode,GatingHierarchy,character,character-method}
\alias{setNode,GatingHierarchy,character,logical-method}
\alias{setNode,GatingSet,character,ANY-method}
\alias{setNode,GatingSet,numeric,ANY-method}
\title{Update the name of one node in a gating hierarchy/GatingSet.}
\usage{
\S4method{setNode}{GatingHierarchy,character,character}(x, y, value, ...)

\S4method{setNode}{GatingHierarchy,character,logical}(x, y, value, ...)

\S4method{setNode}{GatingSet,numeric,ANY}(x, y, value, ...)
}
\arguments{
\item{value}{A \code{character} the name of the node.}

\item{x}{\code{GatingHierarchy} object}

\item{y}{\code{numeric} node index}

\item{value}{\code{logical} whether to hide a node}
}
\description{
\code{setNode} update the name of one node in a gating hierarchy/GatingSet.

hide/unhide a node
}
\examples{
\dontrun{
    #G is a gating hierarchy
    getNodes(G[[1]])#return node names
    setNode(G,"L","lymph")
  }
\dontrun{
     setNode(gh, 4, FALSE) # hide a node
     setNode(gh, 4, TRUE) # unhide a node
}
}

