% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/graphGatingML-methods.R
\docType{methods}
\name{show,graphGML-method}
\alias{show,graphGML-method}
\title{show method for graphGML}
\usage{
\S4method{show}{graphGML}(object)
}
\arguments{
\item{object}{\code{graphGML}}
}
\description{
show method for graphGML
}

