% Generated by roxygen2 (4.0.1): do not edit by hand
\name{subset.GatingSet}
\alias{subset.GatingSet}
\title{subset the GatingSet/GatingSetList based on 'pData'}
\usage{
\method{subset}{GatingSet}(x, ...)
}
\arguments{
\item{x}{\code{GatingSet} or \code{GatingSetList}}

\item{...}{other arguments passed to \link{ncdfFlowSet:subset}}
}
\value{
a code{GatingSet} or \code{GatingSetList} object
}
\description{
subset the GatingSet/GatingSetList based on 'pData'
}

