\name{summary-methods}
\docType{methods}
\alias{summary-methods}
\alias{summary,ANY-method}
\alias{summary,flowJoWorkspace-method}
\alias{show-methods}
\alias{show,GatingHierarchy-method}
\alias{show,GatingSet-method}
\alias{show,flowJoWorkspace-method}
\title{ Summarize a flowJoWorkspace object }
\description{
	Summarize a \code{flowJoWorkspace} object.
}
\keyword{methods}
