% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{use_on_disk_idx}
\alias{use_on_disk_idx}
\title{report/set the on-disk idx flag}
\usage{
use_on_disk_idx(value = NULL)
}
\arguments{
\item{value}{TRUE/FALSE, when not specified, it is NULL and simply report the current setting}
}
\value{
TRUE/FALSE
}
\description{
report/set the on-disk idx flag
}
\examples{
#report the current setting
use_on_disk_idx()
#turn on the ondisk idx
use_on_disk_idx(TRUE)
}
