\name{archive}
\alias{archive}
\alias{unarchive}
\title{archive/unarchive to/from a tar file}
\usage{
  archive(G, file = tempfile())

  unarchive(file, path = tempdir())
}
\arguments{
  \item{G}{a \code{GatingSet}}

  \item{file}{a \code{character} target/source archive file
  name}

  \item{path}{a \code{character} target folder that stores
  cdf file}
}
\description{
  Defunct by save_gs/load_gs
}

