\name{booleanFilter-class}
\alias{booleanFilter-class}
\alias{show}
\alias{show,booleanFilter-method}
\title{A class describing logical operation (& or |) of the reference populations}
\description{
  It inherits class \code{\link{expressionFilter}} and
  exists for the purpose of methods dispatching.
}
\seealso{
  \code{\link{add}} \code{\linkS4class{GatingHierarchy}}
}

