% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{cf_write_h5}
\alias{cf_write_h5}
\title{Save the cytoframe as h5 format}
\usage{
cf_write_h5(cf, filename)
}
\arguments{
\item{cf}{cytoframe object}

\item{filename}{the full path of the output h5 file}
}
\description{
Save the cytoframe as h5 format
}
\seealso{
Other cytoframe/cytoset IO functions: 
\code{\link{cf_get_uri}()},
\code{\link{cf_write_disk}()},
\code{\link{cs_get_uri}()},
\code{\link{load_cytoframe_from_fcs}()},
\code{\link{load_cytoframe}()},
\code{\link{load_cytoset_from_fcs}()}
}
\concept{cytoframe/cytoset IO functions}
