% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{cleanup}
\alias{cleanup}
\alias{cf_cleanup}
\alias{cs_cleanup}
\alias{gh_cleanup}
\alias{gs_cleanup}
\title{Remove on-disk files associatated with flowWorkspace data classes}
\usage{
cf_cleanup(cf)
}
\arguments{
\item{cf}{a cytoframe, cytoset, GatingHierarchy, or GatingSet object}
}
\description{
These methods immediately delete the on-disk  storage associated with \link{cytoframe},
\link{cytoset}, \linkS4class{GatingHierarchy}, or \linkS4class{GatingSet} objects
}
\details{
this will override tempdir() in determining the top directory under which files can safely be removed.
}
