% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gs.R
\name{convert_gs_legacy}
\alias{convert_gs_legacy}
\title{convert the legacy(mixed with R and C++ files) GatingSet archive to the new format(C++ only)}
\usage{
convert_gs_legacy(from, to)
}
\arguments{
\item{from}{the old archive path}

\item{to}{the new archive path}
}
\description{
convert the legacy(mixed with R and C++ files) GatingSet archive to the new format(C++ only)
}
\examples{
\dontrun{
convert_gs_legacy(old_gs_path, new_gs_path)
}
}
