% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoset.R, R/GatingSet_Methods.R
\name{cs_get_uri}
\alias{cs_get_uri}
\alias{cs_get_h5_file_path}
\alias{gs_get_uri}
\title{Return the path of the underlying data files}
\usage{
cs_get_uri(x)

cs_get_h5_file_path(x)

gs_get_uri(x)
}
\description{
Return the path of the underlying data files
}
\seealso{
Other cytoframe/cytoset IO functions: 
\code{\link{cf_get_uri}()},
\code{\link{cf_write_disk}()},
\code{\link{cf_write_h5}()},
\code{\link{load_cytoframe_from_fcs}()},
\code{\link{load_cytoframe}()},
\code{\link{load_cytoset_from_fcs}()}
}
\concept{cytoframe/cytoset IO functions}
