% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{cf_keyword_insert}
\alias{cf_keyword_insert}
\alias{cf_keyword_rename}
\alias{cf_keyword_delete}
\alias{cf_keyword_set}
\title{\code{cytoframe} keyword access methods}
\usage{
cf_keyword_insert(cf, keys, values)

cf_keyword_delete(cf, keys)

cf_keyword_rename(cf, old_keys, new_keys)

cf_keyword_set(cf, keys, values)
}
\arguments{
\item{cf}{a \code{cytoframe}}

\item{keys}{the keyword names to insert/delete/replace -- single value or vector}

\item{values}{the values to associate with the supplied keywords -- single value or vector of sample length as keys}

\item{old_keys}{the old keyword name (for renaming)}

\item{new_keys}{the new keyword name (for renamiing)}
}
\description{
These methods allow for direct insertion, deletion, or renaming
of keywords in \code{\link{cytoframe}} objects.
}
