% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{cf_keyword_insert}
\alias{cf_keyword_insert}
\alias{cf_keyword_rename}
\alias{cf_keyword_delete}
\alias{cf_keyword_set}
\title{\code{cytoframe} keyword access methods}
\usage{
cf_keyword_insert(cf, keyword, value)

cf_keyword_delete(cf, keyword)

cf_keyword_rename(cf, from, to)

cf_keyword_set(cf, values)
}
\arguments{
\item{cf}{a \code{cytoframe}}

\item{keyword}{the keyword name to insert/delete/replace}

\item{value}{the value to associate with the supplied keyword}

\item{from}{the old keyword name (for renaming)}

\item{to}{the new keyword name (for renamiing)}

\item{values}{a named character vector of replacement values, whose names are the keys whose values will be replaced}
}
\description{
These methods allow for direct insertion, deletion, or renaming
of keywords in \code{\link{cytoframe}} objects.
}
