% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterObject_Methods.R
\name{filterObject}
\alias{filterObject}
\alias{filterObject.default}
\alias{filter_to_list}
\alias{filter_to_list.rectangleGate}
\alias{filter_to_list.polygonGate}
\alias{filter_to_list.quadGate}
\alias{filter_to_list.booleanFilter}
\alias{filter_to_list.ellipsoidGate}
\alias{filter_to_list.logical}
\title{convert flowCore filter to a list 
 
It convert the flowCore gate to a list whose structure can be understood by 
underlying c++ data structure.}
\usage{
filterObject(x)

\method{filterObject}{default}(x)

filter_to_list(x)

\method{filter_to_list}{rectangleGate}(x)

\method{filter_to_list}{polygonGate}(x)

\method{filter_to_list}{quadGate}(x)

\method{filter_to_list}{booleanFilter}(x)

\method{filter_to_list}{ellipsoidGate}(x)

\method{filter_to_list}{logical}(x)
}
\arguments{
\item{x}{\code{filter} a flowCore gate. Currently supported gates are:
"rectangleGate", "polygonGate","ellipsoidGate" and "booleanFilter"}
}
\value{
a \code{list}
}
\description{
convert flowCore filter to a list 
 
It convert the flowCore gate to a list whose structure can be understood by 
underlying c++ data structure.
}
