% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterObject-Methods.R
\docType{methods}
\name{filterObject,rectangleGate-method}
\alias{filterObject}
\alias{filterObject,booleanFilter-method}
\alias{filterObject,ellipsoidGate-method}
\alias{filterObject,polygonGate-method}
\alias{filterObject,rectangleGate-method}
\title{convert flowCore filter to a list}
\usage{
\S4method{filterObject}{rectangleGate}(x)

\S4method{filterObject}{polygonGate}(x)

\S4method{filterObject}{booleanFilter}(x)

\S4method{filterObject}{ellipsoidGate}(x)
}
\arguments{
\item{x}{\code{filter} a flowCore gate. Currently supported gates are:
"rectangleGate", "polygonGate","ellipsoidGate" and "booleanFilter"}
}
\value{
a \code{list}
}
\description{
It convert the flowCore gate to a list whose structure can be understood by
underlying c++ data structure.
}

