% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{flowJoWorkspace-class}
\alias{flowJoWorkspace-class}
\alias{show,flowJoWorkspace-method}
\title{An R representation of a flowJo workspace.}
\description{
Objects can be created by calls of the form \code{new("flowJoWorkspace.xml", ...)}.
}
\section{Slots}{

\describe{
  \item{\code{version}:}{Object of class \code{"character"}. The version of the XML workspace. }
  \item{\code{file}:}{Object of class \code{"character"}. The file name. }
  \item{\code{.cache}:}{Object of class \code{"environment"}. An environment for internal use.  }
	\item{\code{path}:}{Object of class \code{"character"}. The path to the file. }
  \item{\code{doc}:}{Object of class \code{"XMLInternalDocument"}. The XML document object. }
  \item{\code{options}:}{Object of class \code{"integer"}. The XML parsing options passed to \code{\link{xmlTreeParse}}. }
  }
}
\examples{
require(flowWorkspaceData)
  d<-system.file("extdata",package="flowWorkspaceData")
  wsfile<-list.files(d,pattern="A2004Analysis.xml",full=TRUE)
  ws <- openWorkspace(wsfile);
  ws
  getSamples(ws)
}
\seealso{
\code{\linkS4class{GatingSet}}
  \code{\linkS4class{GatingHierarchy}}
}

