% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/GatingHierarchy_Methods.R,
%   R/GatingSet_Methods.R, R/GatingSetList_Methods.R, R/add_Methods.R,
%   R/copyNode.R, R/deprecated.R, R/flow_trans.R, R/getDescendants.R,
%   R/getStats.R, R/merge_GatingSet.R, R/moveNode.R, R/updateIndices.R,
%   R/utils.R
\name{flowWorkspace-deprecated}
\alias{flowWorkspace-deprecated}
\title{Deprecated functions in package \pkg{flowWorkspace}.}
\description{
\code{getNodes} --> \code{\link{gs(/gh)_get_nodes}}

\code{setNode} --> \code{\link{gs(/gh)_set_node/gs(/gh)_hide_node/gs(/gh)_unhide_node}}

\code{getData} --> \code{\link{gs(/gh)_cyto_data}}

\code{flowData} --> \code{\link{gs_cyto_data}}

\code{flowData<-} --> \code{\link{gs_cyto_data<-}}

\code{filterObject} --> \code{\link{filter_to_list}}

\code{plotGate} --> \code{\link{autoplot}}

\code{getPopStats} --> \code{\link{gs(/gh)_get_popstats}}

\code{plotPopCV} --> \code{\link{gh_plot_popcv}}

\code{getGate} --> \code{\link{gs(/gh)_get_gate}}

\code{setGate} --> \code{\link{gs(/gh)_set_gate}}

\code{getParent} --> \code{\link{gs(/gh)_get_parent}}

\code{getChildren} --> \code{\link{gs(/gh)_get_children}}

\code{getProp} --> \code{\link{gh_get_proportion}}

\code{getTotal} --> \code{\link{gs(/gh)_get_count}}

\code{getSingleCellExpression} --> \code{\link{gs_get_singlecell_expression}}

\code{getFullNodePath} --> \code{\link{gh_get_full_path}}

\code{openWorkspace} --> \code{\link{open_workspace}}

\code{isGated} --> \code{\link{gh_is_gated}}

\code{isNegated} --> \code{\link{gh_is_negated}}

\code{isHidden} --> \code{\link{gh_is_hidden}}

\code{openWorkspace} --> \code{\link{open_workspace}}

\code{getTransformations} --> \code{\link{gh_get_transformations}}

\code{getCompensationMatrices} --> \code{\link{gh_get_compensations}}

\code{isNcdf} --> \code{\link{gs_is_h5}}

\code{clone} --> \code{\link{gs_clone}}

\code{recompute} --> \code{\link{gs_recompute}}

\code{getLoglevel} --> \code{\link{get_log_level}}

\code{setLoglevel} --> \code{\link{set_log_level}}

\code{rbind2} --> \code{\link{gslist_to_gs}}

\code{add} --> \code{\link{gs(/gh)_add_gate}}

\code{Rm} --> \code{\link{gs(/gh)_remove_gate}}

\code{copyNode} --> \code{\link{gh_copy_gate}}

\code{openWorkspace} --> \code{\link{open_workspace}}

\code{flowJo.flog} --> \code{\link{flowjo_flog}}

\code{flowJoTrans} --> \code{\link{flowjo_biexp}}

\code{flowJo.fasinh} --> \code{\link{flowjo_fasinh}}

\code{flowJo.fsinh} --> \code{\link{flowjo_fsinh}}

\code{flowJo_fasinh_trans} --> \code{\link{flowjo_fasinh_trans}}

\code{getDescendants} --> \code{\link{gh_get_descendants}}

\code{getStats} --> \code{\link{gs(/gh)_get_stats}}

\code{groupByTree} --> \code{\link{gs_group_by_tree}}

\code{groupByChannels} --> \code{\link{gs_group_by_channels}}

\code{checkRedundantNodes} --> \code{\link{gs_check_redundant_nodes}}

\code{dropRedundantNodes} --> \code{\link{gs_drop_redundant_nodes}}

\code{dropRedundantChannels} --> \code{\link{gs_drop_redundant_channels}}

\code{updateChannels} --> \code{\link{gs_update_channels}}

\code{moveNode} --> \code{\link{gh_move_node}}

\code{updateIndices} --> \code{\link{gh_set_pop_indices}}

\code{getMergedStats} --> \code{\link{gs_get_merged_stats}}

\code{set.count.xml} --> \code{\link{gh_set_xml_count}}
}
