% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R, R/GatingHierarchy_Methods.R,
%   R/GatingSet_Methods.R, R/GatingSetList_Methods.R, R/filterObject_Methods.R,
%   R/add_Methods.R, R/copyNode.R, R/deprecated.R, R/flow_trans.R,
%   R/getDescendants.R, R/getSingleCellExpression.R, R/merge_GatingSet.R,
%   R/setGate_Methods.R, R/updateIndices.R, R/utils.R
\name{flowWorkspace-deprecated}
\alias{flowWorkspace-deprecated}
\title{Deprecated functions in package \pkg{flowWorkspace}.}
\description{
\code{getStats} --> \code{gs(/gh)_pop_get_stats}

\code{getProp} --> \code{gh_pop_get_proportion}

\code{getTotal} --> \code{gh_pop_get_count}

\code{getPopStats} --> \code{gs(/gh)_pop_get_stats}

\code{getNodes} --> \code{gs_get_pop_paths}

\code{getParent} --> \code{gs_pop_get_parent}

\code{getChildren} --> \code{gs_pop_get_children}

\code{getGate} --> \code{gs(/gh)_get_gate}

\code{getIndices} --> \code{gh_pop_get_indices}

\code{isGated} --> \code{gh_pop_is_gated}

\code{isNegated} --> \code{gh_pop_is_negated}

\code{isHidden} --> \code{gh_pop_is_hidden}

\code{getData} --> \code{gs(/gh)_get_data}

\code{getTransformations} --> \code{gh_get_transformations}

\code{getCompensationMatrices} --> \code{gh_get_compensations}

\code{setNode} --> \code{gs(/gh)_set_node_name/gs(/gh)_set_node_visible}

\code{isNcdf} --> \code{gs_is_h5}

\code{flowData} --> \code{gs_cyto_data}

\code{flowData<-} --> \code{gs_cyto_data<-}

\code{getLoglevel} --> \code{get_log_level}

\code{setLoglevel} --> \code{set_log_level}

\code{rbind2} --> \code{gslist_to_gs}

\code{filterObject} --> \code{filter_to_list}

\code{add} --> \code{gs_pop_add}

\code{Rm} --> \code{gs_pop_remove}

\code{copyNode} --> \code{gh_copy_gate}

\code{openWorkspace} --> \code{open_flowjo_xml}

\code{flowJo.flog} --> \code{flowjo_log_trans}

\code{flowJoTrans} --> \code{flowjo_biexp}

\code{flowJo_biexp_trans} --> \code{flowjo_biexp_trans}

\code{flowJo.fasinh} --> \code{flowjo_fasinh}

\code{flowJo.fsinh} --> \code{flowjo_fsinh}

\code{flowJo_fasinh_trans} --> \code{flowjo_fasinh_trans}

\code{getDescendants} --> \code{gh_pop_get_descendants}

\code{getSingleCellExpression} --> \code{gs_get_singlecell_expression}

\code{groupByTree} --> \code{gs_split_by_tree}

\code{groupByChannels} --> \code{gs_split_by_channels}

\code{checkRedundantNodes} --> \code{gs_check_redundant_nodes}

\code{dropRedundantNodes} --> \code{gs_remove_redundant_nodes}

\code{dropRedundantChannels} --> \code{gs_drop_redundant_channels}

\code{updateChannels} --> \code{gs_update_channels}

\code{setGate} --> \code{gs(/gh)_pop_set_gate}

\code{updateIndices} --> \code{gh_pop_set_indices}

\code{getMergedStats} --> \code{gs_pop_get_count_with_meta}

\code{set.count.xml} --> \code{gh_pop_set_xml_count}
}
