% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{package}
\name{flowWorkspace-package}
\alias{flowWorkspace}
\alias{flowWorkspace-package}
\title{Import and replicate flowJo workspaces and gating schemes using flowCore.}
\description{
Import flowJo workspaces into R. Generate the flowJo gating hierarchy and gates using flowCore functionality. Transform and compensate data in accordance with flowJo settings. Plot gates, gating hierarchies,  population statistics, and compare flowJo vs flowCore population summaries.
}
\details{
\tabular{ll}{
    Package: \tab flowWorkspace\cr
    Type: \tab Package\cr
    Version: \tab 0.5.40\cr
    Date: \tab 2011-03-04\cr
    License: \tab Artistic 2.0 \cr
    LazyLoad: \tab yes\cr
    Depends: \tab R (>= 2.16.0),Rcpp (>= 0.9.9)\cr
  }
}
\author{
Greg Finak, Mike Jiang
}
\references{
\url{http://www.rglab.org/}
}

