% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz.R
\name{flowWorkspace.par.init}
\alias{flowWorkspace.par.init}
\title{workspace version is parsed from xml node '/Workspace/version' in flowJo workspace
and matched with this list to dispatch to the one of the three workspace parsers}
\usage{
flowWorkspace.par.init()
}
\description{
workspace version is parsed from xml node '/Workspace/version' in flowJo workspace
and matched with this list to dispatch to the one of the three workspace parsers
}

