% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/GatingSet_Methods.R
\docType{methods}
\name{GatingSet,cytoSet-method}
\alias{GatingSet,cytoSet-method}
\alias{plotGate,GatingSet,missing-method}
\alias{plotGate}
\alias{plotGate-methods}
\alias{plotGate,GatingHierarchy,character-method}
\alias{plotGate,GatingHierarchy,numeric-method}
\alias{plotGate,GatingHierarchy,missing-method}
\alias{plotGate,GatingSet,numeric-method}
\alias{plotGate,GatingSet,character-method}
\title{constructors for GatingSet}
\usage{
\S4method{GatingSet}{cytoSet}(x)

\S4method{plotGate}{GatingSet,missing}(x, y, ...)
}
\arguments{
\item{x}{\code{\linkS4class{GatingSet}} or \code{\linkS4class{GatingHierarchy}}object}

\item{y}{\code{character} the node name or full(/partial) gating path
or \code{numeric} representing the node index in the \code{GatingHierarchy}.
or \code{missing} which will plot all gates and one gate per page. It is useful for generating plots in a multi-page pdf.
Nodes can be accessed with \code{\link{getNodes}}.}

\item{...}{other arguments. see \link{parseWorkspace}}

\item{path}{\code{character} specifies the path to the flow data (FCS files)}

\item{...}{\itemize{
 \item{bool}{ \code{logical} specifying whether to plot boolean gates.}
 \item{arrange.main}{ \code{character} The title of the main page of the plot. Default is the sample name. Only valid when \code{x} is GatingHierarchy}
 \item{arrange}{ \code{logical} indicating whether to arrange different populations/nodes on the same page via \code{arrangeGrob} call.}
 \item{merge}{ \code{logical} indicating whether to draw multiple gates on the same plot if these gates share the same parent population and same x,y dimensions/parameters;}
\item{projections}{ \code{list} of character vectors used to customize x,y axis. By default, the x,y axis are determined by the respective gate parameters.
                                The elements of the list are named by the population name or path (see \code{y}). Each element is a pair of named character specifying the channel name(or marker name) for x, y axis.
                                Short form of channel or marker names (e.g. "APC" or "CD3") can be used as long as they can be uniquely matched to the dimentions of flow data.
                                For example, projections = list("lymph" = c(x = "SSC-A", y = "FSC-A"), "CD3" = c(x = "CD3", y = "SSC-A"))
                     }
\item{par.settings}{ \code{list} of graphical parameters passed to \code{\link{lattice}};}

 \item{gpar}{ \code{list} of grid parameters passed to \code{\link{grid.layout}};}

 \item{lattice}{ \code{logical} deprecated;}

 \item{formula}{ \code{formula} a formula passed to \code{xyplot} function of \code{flowViz}, by default it is NULL, which means the formula is generated according to the x,y parameters associated with gate.}

 \item{cond}{ \code{character} the conditioning variable to be passed to lattice plot.}

 \item{overlay}{Node names. These populations are plotted on top of the existing gates(defined by \code{y} argument) as the overlaid dots.}
 \item{overlay.symbol}{A named (lattice graphic parameter) list that defines the symbol color and size for each overlaid population.
                        If not given, we automatically assign the colors.}
 \item{key}{Lattice legend paraemter for overlay symbols.}

 \item{default.y}{ \code{character} specifiying y channel for xyplot when plotting a 1d gate. Default is "SSC-A" and session-wise setting can be stored by 'flowWorkspace.par.set("plotGate", list(default.y = "FSC-A"))'}

 \item{type}{ \code{character} either "xyplot" or "densityplot". Default is "xyplot"  and session-wise setting can be stored by 'flowWorkspace.par.set("plotGate", list(type = "xyplot"))'}

 \item{fitGate}{ used to disable behavior of plotting the gate region in 1d densityplot. Default is FALSE and  session-wise setting can be stored by 'flowWorkspace.par.set("plotGate", list(fitGate = FALSE))'}

 \item{strip}{ \code{ligcal} specifies whether to show pop name in strip box,only valid when x is \code{GatingHierarchy}}
\item{strip.text}{either "parent" (the parent population name) or "gate "(the gate name).}


 \item{raw.scale}{ \code{logical} whether to show the axis in raw(untransformed) scale. Default is TRUE and can be stored as session-wise setting by 'flowWorkspace.par.set("plotGate", list(raw.scale = TRUE))'}
 \item{xlim, ylim}{ \code{character} can be either "instrument" or "data" which determines the x, y axis scale
                                           either by instrument measurement range or the actual data range.
                    or \code{numeric} which specifies customized range.
                     They can be stored as session-wise setting by 'flowWorkspace.par.set("plotGate", list(xlim = "instrument"))'
}

 \item{...}{

         path A \code{character} or \code{numeric} scalar passed to \link{getNodes} method (used to control how the gating/node path is displayed)

         ... The other additional arguments to be passed to \link[flowViz]{xyplot}.
         }
}}
}
\value{
a \code{trellis} object if \code{arrange} is \code{FALSE},
}
\description{
construct a gatingset with empty trees (just root node)

construct object from existing gating hierarchy(gating template) and flow data
}
\examples{
\dontrun{
#fdata could be a flowSet or ncdfFlowSet
gs <- GatingSet(fdata)
}
\dontrun{
	#G is a GatingHierarchy
	plotGate(G,getNodes(G)[5]);#plot the gate for the  fifth node
}
}
\references{
\url{http://www.rglab.org/}
}
