% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{GatingSet}
\alias{GatingSet}
\alias{GatingSet,GatingHierarchy,character-method}
\alias{GatingSet,character,character-method}
\alias{GatingSet,flowSet,ANY-method}
\title{constructors for GatingSet}
\usage{
GatingSet(x, y, ...)

\S4method{GatingSet}{character,character}(x, y, includeGates = FALSE,
  sampNloc = "keyword", xmlParserOption, wsType, ...)

\S4method{GatingSet}{flowSet,ANY}(x, y, ...)

\S4method{GatingSet}{GatingHierarchy,character}(x, y, path = ".",
  isNcdf = FALSE, ...)
}
\description{
construct object from xml workspace file and a list of sampleIDs

construct a gatingset with empty trees (just root node)

construct object from existing gating hierarchy(gating template) and flow data
}

