% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/GatingSet_Methods.R
\docType{methods}
\name{GatingSet,cytoset,ANY-method}
\alias{GatingSet,cytoset,ANY-method}
\alias{GatingSet,GatingHierarchy,character-method}
\title{constructors for GatingSet}
\usage{
\S4method{GatingSet}{cytoset,ANY}(x)

\S4method{GatingSet}{GatingHierarchy,character}(x, y, path = ".", ...)
}
\arguments{
\item{path}{\code{character} specifies the path to the flow data (FCS files)}

\item{...}{other arguments. see \link{parseWorkspace}}
}
\description{
construct a gatingset with empty trees (just root node)

construct object from existing gating hierarchy(gating template) and flow data
}
\examples{
\dontrun{
#fdata could be a flowSet or ncdfFlowSet
gs <- GatingSet(fdata)
}
}
