% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{methods}
\name{GatingSet,cytoset,ANY-method}
\alias{GatingSet,cytoset,ANY-method}
\title{constructors for GatingSet}
\usage{
\S4method{GatingSet}{cytoset,ANY}(x)
}
\description{
construct a gatingset with empty trees (just root node)
}
\examples{
\dontrun{
#fdata could be a flowSet or ncdfFlowSet
gs <- GatingSet(fdata)
}
}
