% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{GatingSet,character,character-method}
\alias{GatingSet}
\alias{GatingSet,GatingHierarchy,character-method}
\alias{GatingSet,character,character-method}
\alias{GatingSet,flowSet,ANY-method}
\title{constructors for GatingSet}
\usage{
\S4method{GatingSet}{character,character}(x, y, includeGates = FALSE,
  sampNloc = "keyword", xmlParserOption, wsType)

\S4method{GatingSet}{flowSet,ANY}(x)

\S4method{GatingSet}{GatingHierarchy,character}(x, y, path = ".",
  isNcdf = FALSE, ...)
}
\arguments{
\item{x}{\code{character} or \code{flowSet} or \code{GatingHierarchy}}

\item{y}{\code{character} or\code{missing}}

\item{includeGates}{\code{logical} whether to parse the gates or just simply extract the flowJo stats}

\item{sampNloc}{\code{character} scalar indicating where to get sampleName(or FCS filename) within xml workspace. It is either from "keyword" or "sampleNode".}

\item{xmlParserOption}{\code{integer} option passed to \code{\link{xmlTreeParse}}}

\item{wsType}{\code{character} workspace type, can be value of "win", "macII", "vX", "macIII".}

\item{path}{\code{character} specifies the path to the flow data (FCS files)}

\item{isNcdf}{\code{logical} whether to use ncdfFlowSet or flowSet as the underlying flow data storage}

\item{...}{other arguments. see \link{parseWorkspace}}
}
\description{
construct object from xml workspace file and a list of sampleIDs (not intended to be called by user.)

construct a gatingset with empty trees (just root node)

construct object from existing gating hierarchy(gating template) and flow data
}
\examples{
\dontrun{
#fdata could be a flowSet or ncdfFlowSet
gs <- GatingSet(fdata)
}
}

