% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R, R/GatingSet_Methods.R,
%   R/identifier.R
\docType{methods}
\name{GatingSet,cytoset,ANY-method}
\alias{GatingSet,cytoset,ANY-method}
\alias{GatingSet,GatingHierarchy,character-method}
\alias{identifier,GatingSet-method}
\alias{identifier,GatingSetList-method}
\alias{identifier<-,GatingSet,ANY-method}
\alias{identifier<-,GatingSetList,character-method}
\title{constructors for GatingSet}
\usage{
\S4method{GatingSet}{cytoset,ANY}(x)

\S4method{GatingSet}{GatingHierarchy,character}(x, y, path = ".",
  swap_cols = FALSE, ...)

\S4method{identifier}{GatingSet}(object)

\S4method{identifier}{GatingSetList}(object)

\S4method{identifier}{GatingSet,ANY}(object) <- value

\S4method{identifier}{GatingSetList,character}(object) <- value
}
\arguments{
\item{x}{GatingSet}

\item{y}{GatingHierarchy}

\item{path}{\code{character} specifies the path to the flow data (FCS files)}

\item{...}{other arguments.}

\item{object}{GatingSet}

\item{value}{string}
}
\description{
construct a gatingset with empty trees (just root node)

construct object from existing gating hierarchy(gating template) and flow data
}
\examples{
\dontrun{
#fdata could be a flowSet or ncdfFlowSet
gs <- GatingSet(fdata)
}
}
