% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{GatingSet-methods}
\alias{GatingSet-methods}
\alias{GatingSet,cytoset,ANY-method}
\alias{GatingSet,flowSet-method}
\title{constructors for GatingSet}
\usage{
GatingSet(x)
}
\arguments{
\item{x}{a flowSet, ncdfFlowSet, or cytoset}
}
\description{
construct a gatingset with empty trees (just root node)
}
\examples{
\dontrun{
#fdata could be a flowSet, ncdfFlowSet, or GatingSet
gs <- GatingSet(fdata)
}
}
