\docType{methods}
\name{GatingSet}
\alias{GatingSet}
\alias{GatingSet,character,character-method}
\alias{GatingSet,flowSet,ANY-method}
\alias{GatingSet,GatingHierarchy,character-method}
\title{constructors for GatingSet}
\description{
  construct object from xml workspace file and a list of
  sampleIDs

  construct a gatingset with empty trees (just root node)

  construct object from existing gating hierarchy(gating
  template) and flow data
}

