% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{getCompensationMatrices,GatingHierarchy-method}
\alias{getCompensationMatrices}
\alias{getCompensationMatrices,GatingHierarchy-method}
\title{Retrieve the compensation matrices from a GatingHierarchy}
\usage{
\S4method{getCompensationMatrices}{GatingHierarchy}(x)
}
\arguments{
\item{x}{A \code{GatingHierarchy} object.}
}
\value{
A list of \code{matrix} representing the spillover matrix in \code{GatingHierarchy}
}
\description{
Retrieve the compensation matrices from a GatingHierarchy.
}
\details{
Return all the compensation matrices in a GatingHierarchy.
}
\examples{
\dontrun{
	#Assume gh is a GatingHierarchy
  getCompensationMatrices(gh);
}
}

