% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R,
%   R/GatingSet_Methods.R
\name{getCompensationMatrices}
\alias{getCompensationMatrices}
\alias{getCompensationMatrices.GatingHierarchy}
\alias{getCompensationMatrices.GatingSet}
\title{Retrieve the compensation matrices from a GatingHierarchy}
\usage{
getCompensationMatrices(x)

\method{getCompensationMatrices}{GatingHierarchy}(x)

\method{getCompensationMatrices}{GatingSet}(x)
}
\arguments{
\item{x}{A \code{GatingHierarchy} object.}
}
\value{
A list of \code{matrix} representing the spillover matrix in \code{GatingHierarchy}
}
\description{
Retrieve the compensation matrices from a GatingHierarchy.
}
\details{
Return all the compensation matrices in a GatingHierarchy.
}
\examples{
  \dontrun{
	#Assume gh is a GatingHierarchy
  getCompensationMatrices(gh);
}
}
