% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingHierarchy-Methods.R, R/GatingSet-Methods.R, R/GatingSetList-Methods.R
\docType{methods}
\name{getData,GatingHierarchy,missing-method}
\alias{getData}
\alias{getData,GatingHierarchy,character-method}
\alias{getData,GatingHierarchy,missing-method}
\alias{getData,GatingSet,character-method}
\alias{getData,GatingSet,missing-method}
\alias{getData,GatingSetList,ANY-method}
\title{get gated flow data from a GatingHierarchy/GatingSet/GatingSetList}
\usage{
\S4method{getData}{GatingHierarchy,missing}(obj, y, ...)

\S4method{getData}{GatingHierarchy,character}(obj, y, ...)

\S4method{getData}{GatingSet,missing}(obj, y, ...)

\S4method{getData}{GatingSet,character}(obj, y, ...)

\S4method{getData}{GatingSetList,ANY}(obj, y, ...)
}
\arguments{
\item{obj}{A \code{GatingHierarchy}, \code{GatingSet} or \code{GatingSetList} object.}

\item{y}{\code{character}  the node name or full(/partial) gating path or \code{numeric} node index.
If not specified, will return the complete flowFrame/flowSet at the root node.}

\item{...}{arguments passed to ncdfFlow::[[}
}
\value{
A \code{flowFrame} object if \code{obj} is a GatingHierarchy.
A \code{flowSet} or \code{ncdfFlowSet} if a \code{GatingSet}.
A \code{ncdfFlowList} if a \code{GatingSetList}.
}
\description{
get gated flow data from a GatingHierarchy/GatingSet/GatingSetList
}
\details{
Returns a flowFrame/flowSet containing the events in the gate defined at node \code{y}.
Subset membership can be obtained using \code{getIndices}.
Population statistics can be obtained using \code{getPop} and \code{getPopStats}.
When calling \code{getData} on a GatingSet,the trees representing the GatingHierarchy for each sample in the GaingSet are presumed to have the same structure.
To update the data, use \code{flowData} method.
}
\examples{
\dontrun{
    #G is a GatingSet
    geData(G,3) #get a flowSet constructed from the third node / population in the tree.
    geData(G,"cd4")

    #gh is a GatingHierarchy
    getData(gh)
}
}
\seealso{
\code{\link{flowData}} \code{\link{getIndices}} \code{\link{getPopStats}}
}

