% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet-Methods.R
\name{getIndiceMat}
\alias{getIndiceMat}
\title{Return the single-cell matrix of 1/0 dichotomized expression}
\usage{
getIndiceMat(gh, y)
}
\arguments{
\item{gh}{\code{GatingHierarchy} object}

\item{y}{\code{character} node name}
}
\description{
Return the single-cell matrix of 1/0 dichotomized expression
}

