% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{getKeywords,flowJoWorkspace,character-method}
\alias{getKeywords}
\alias{getKeywords,flowJoWorkspace,character-method}
\title{Get Keywords}
\usage{
\S4method{getKeywords}{flowJoWorkspace,character}(obj, y)
}
\arguments{
\item{obj}{A \code{flowJoWorkspace}}

\item{y}{c\code{character} specifying the sample names}
}
\value{
A list of keyword - value pairs.
}
\description{
Retrieve keywords associated with a workspace
}
\details{
Retrieve a list of keywords from a \code{flowJoWorkspace}
}
\examples{
require(flowWorkspaceData)
  d<-system.file("extdata",package="flowWorkspaceData")
  wsfile<-list.files(d,pattern="manual.xml",full=TRUE)
  ws <- openWorkspace(wsfile);

  getSamples(ws)
  getKeywords(ws,"CytoTrol_CytoTrol_1.fcs")
}

