% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowJoWorkspace-Methods.R
\docType{methods}
\name{getKeywords,flowJoWorkspace,character-method}
\alias{getKeywords}
\alias{getKeywords,flowJoWorkspace,character-method}
\alias{getKeywords,flowJoWorkspace,numeric-method}
\title{Get Keywords}
\usage{
\S4method{getKeywords}{flowJoWorkspace,character}(obj, y, ...)

\S4method{getKeywords}{flowJoWorkspace,numeric}(obj, y, ...)
}
\arguments{
\item{obj}{A \code{flowJoWorkspace}}

\item{y}{c\code{character} or \code{numeric} specifying the sample name or sample ID}

\item{...}{other arguments
     sampNloc a \code{character} the location where the sample name is specified. See \code{parseWorkspace} for more details.}
}
\value{
A list of keyword - value pairs.
}
\description{
Retrieve keywords associated with a workspace
}
\details{
Retrieve a list of keywords from a \code{flowJoWorkspace}
}
\examples{
require(flowWorkspaceData)
  d<-system.file("extdata",package="flowWorkspaceData")
  wsfile<-list.files(d,pattern="manual.xml",full=TRUE)
  ws <- openWorkspace(wsfile);

  getSamples(ws)
  res <- try(getKeywords(ws,"CytoTrol_CytoTrol_1.fcs"), silent = TRUE)
  print(res[[1]])
  getKeywords(ws, 1)
}

