% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingHierarchy-Methods.R
\docType{methods}
\name{getNodes,GatingSet-method}
\alias{getNodes}
\alias{getNodes,GatingSet-method}
\title{Get the names of all nodes from a gating hierarchy.}
\usage{
\S4method{getNodes}{GatingSet}(x, y = NULL, order = "regular",
  path = "full", showHidden = FALSE, ...)
}
\arguments{
\item{x}{A \code{GatingSet} Assuming the gating hierarchy are identical within the \code{GatingSet}, the Gating tree of the first sample is used to query the node information.}

\item{y}{A \code{character} not used.}

\item{order}{\code{order=c("regular","tsort","bfs")} returns the nodes in regular, topological or breadth-first sort order.
"regular" is default.}

\item{path}{A \code{character} or \code{numeric} scalar. when \code{numeric}, it specifies the fixed length of gating path (length 1 displays terminal name).
When \code{character}, it can be either 'full' (full path, which is default) or 'auto' (display the shortest unique gating path from the bottom of gating tree).}

\item{showHidden}{\code{logical} whether to include the hidden nodes}

\item{...}{Additional arguments.}
}
\value{
getNodes returns a \code{character} vector of node/population names, ordered appropriately.
}
\description{
\code{getNodes} returns a character vector of names of the nodes (populations) in the \code{GatingSet}.
}
\details{
integer indices of nodes are based on regular order,so whenver need to map from character node name to integer node ID,make sure
to use default order which is regular.
}
\examples{
\dontrun{
    #G is a gating hierarchy
    getNodes(G, path = 1])#return node names (without prefix)
    getNodes(G,path = "full")#return the full path
    getNodes(G,path = 2)#return the path as length of two
    getNodes(G,path = "auto)#automatically determine the length of path
    setNode(G,"L","lymph")
  }
}

