% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowJoWorkspace_Methods.R
\name{getSamples.flowJoWorkspace}
\alias{getSamples.flowJoWorkspace}
\alias{getSamples}
\title{Get a list of samples from a flowJo workspace}
\usage{
\method{getSamples}{flowJoWorkspace}(x, group_id = NULL)
}
\arguments{
\item{x}{A \code{flowJoWorkspace}}

\item{group_id}{\code{integer} specifies the group from which samples are returned}
}
\value{
A \code{data.frame} with columns \code{sampleID}, \code{name}
}
\description{
Return  a data frame of samples contained in a flowJo workspace
}
\details{
The samples with 0 populations are excluded.
Returns a \code{data.frame} of samples in the \code{flowJoWorkspace}, including their
\code{sampleID}, \code{name}
}
\examples{
      \dontrun{
        #ws is a flowJoWorkspace
        getSamples(ws);
      }
}
