% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowJoWorkspace-Methods.R
\docType{methods}
\name{getSamples,flowJoWorkspace-method}
\alias{getSamples}
\alias{getSamples,flowJoWorkspace-method}
\title{Get a list of samples from a flowJo workspace}
\usage{
\S4method{getSamples}{flowJoWorkspace}(x, sampNloc = "keyword")
}
\arguments{
\item{x}{A \code{flowJoWorkspace}}

\item{sampNloc}{\code{character} either "keyword" or "sampleNode". see \link{parseWorkspace}}
}
\value{
A \code{data.frame} with columns \code{sampleID}, \code{name}, and \code{compID} if \code{x} is a \code{flowJoWorkspace}.
}
\description{
Return  a data frame of samples contained in a flowJo workspace
}
\details{
Returns a \code{data.frame} of samples in the \code{flowJoWorkspace}, including their
\code{sampleID}, \code{name}, and \code{compID} (compensation matrix ID).
}
\examples{
\dontrun{
        #ws is a flowJoWorkspace
        getSamples(ws);
      }
}

