% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\docType{methods}
\name{GatingSet,GatingHierarchy,character-method}
\alias{GatingSet,GatingHierarchy,character-method}
\alias{gh_apply_to_new_fcs}
\title{constructors for GatingSet}
\usage{
\S4method{GatingSet}{GatingHierarchy,character}(x, y, path = ".", ...)

gh_apply_to_new_fcs(x, files, ...)
}
\arguments{
\item{x}{GatingHierarchy}

\item{y}{sample names}

\item{path}{\code{character} specifies the path to the flow data (FCS files)}

\item{...}{other arguments.}

\item{files}{fcs file paths}

\item{swap_cols}{for internal usage}
}
\description{
construct object from existing gating hierarchy(gating template) and flow data
}
