% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R,
%   R/GatingSet_Methods.R
\name{gh_get_compensations}
\alias{gh_get_compensations}
\alias{getCompensationMatrices}
\alias{getCompensationMatrices,GatingHierachy-method}
\alias{gs_get_compensations}
\title{Retrieve the compensation matrices from a \code{GatingHierarchy} or \code{GatingSet}}
\usage{
gh_get_compensations(x)

gs_get_compensations(x)
}
\arguments{
\item{x}{A \code{GatingHierarchy} or \code{GatingSet} object.}
}
\value{
A list of \code{matrix} representing the spillover matrix in \code{GatingHierarchy} or \code{GatingSet}
}
\description{
Retrieve the compensation matrices from a \code{GatingHierarchy} or \code{GatingSet}.
}
\details{
Return all the compensation matrices in a \code{GatingHierarchy} or \code{GatingSet}
}
\examples{
  \dontrun{
	 # Assume gh is a GatingHierarchy and gs is a GatingSet
  gh_get_compensations(gh)
  gs_get_compensations(gs)
}
}
