% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{gh_pop_compare_stats}
\alias{gh_pop_compare_stats}
\alias{gh_pop_stats_compare}
\alias{getPopStats}
\alias{getPopStats,GatingHierarchy-method}
\alias{gh_pop_stats_ls}
\title{Compare the stats(count/freq) between the version parsed from xml and the one recalculated/gated from R}
\usage{
gh_pop_stats_compare(
  x,
  nodes = NULL,
  path = "auto",
  type = c("Count"),
  legacy = FALSE,
  ...
)

gh_pop_compare_stats(...)

gh_pop_stats_ls(x, node = NULL, ...)
}
\arguments{
\item{x}{GatingHierarchy}

\item{path}{see \link{gs_get_pop_paths}}

\item{...}{not used}

\item{node}{the population node to retrieve the stats from. Default is NUll, meaning retrieving stats type all nodes.}
}
\description{
Compare the stats(count/freq) between the version parsed from xml and the one recalculated/gated from R

list the stats type available in the gating tree
}
