% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getStats.R
\name{gh_pop_compare_stats}
\alias{gh_pop_compare_stats}
\alias{getPopStats}
\alias{getPopStats,GatingHierarchy-method}
\title{Compare the stats(count/freq) between the version parsed from xml and the one recalculated/gated from R}
\usage{
gh_pop_compare_stats(x, path = "auto", ...)
}
\arguments{
\item{x}{GatingHierarchy}

\item{path}{see \link{gs_get_pop_paths}}

\item{...}{not used}
}
\description{
Compare the stats(count/freq) between the version parsed from xml and the one recalculated/gated from R
}
