% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gs_get_leaf_nodes}
\alias{gs_get_leaf_nodes}
\alias{gh_get_leaf_nodes}
\alias{get_leaf_nodes}
\title{get all the leaf nodes}
\usage{
gs_get_leaf_nodes(x, ancestor = "root", ...)

gh_get_leaf_nodes(x, ancestor = "root", ...)
}
\arguments{
\item{x}{GatingHierarchy/GatingSet object}

\item{ancestor}{ancestor node where the leaf nodes descend from. Default is 'root'.}

\item{...}{arguments passed to 'gs_get_pop_paths" method}
}
\value{
the leaf nodes
}
\description{
get all the leaf nodes
}
