% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R,
%   R/GatingSet_Methods.R
\name{gs_pop_set_name}
\alias{gs_pop_set_name}
\alias{gh_pop_set_name}
\alias{setNode}
\alias{setNode,GatingHierarchy,character,character-method}
\alias{setNode,GatingHierarchy,character,ANY-method}
\alias{setNode,GatingSet,character,ANY-method}
\title{Update the name of one node in a gating hierarchy/GatingSet.}
\usage{
gh_pop_set_name(x, y, value)

gs_pop_set_name(x, y, value)
}
\arguments{
\item{x}{GatingHierarchy}

\item{y}{pop name/path}

\item{value}{A \code{character} the name of the node}
}
\description{
\code{gh_pop_set_name/gs_pop_set_name} update the name of one node in a gating hierarchy/GatingSet.
}
\examples{
\dontrun{
    # G is a GatingHierarchy
    gs_get_pop_paths(G[[1]])#return node names
    gh_pop_set_name(G,"L","lymph")
}
}
