% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSetList_Methods.R
\name{gslist_to_gs}
\alias{gslist_to_gs}
\alias{rbind2,GatingSetList,missing-method}
\title{Merge a GatingSetList into a single GatingSet}
\usage{
gslist_to_gs(x, ...)
}
\arguments{
\item{x}{GatingSetList}

\item{...}{other arguments passed to \code{gslist_to_gs} method for \code{ncdfFlowList}}
}
\description{
Merge a GatingSetList into a single GatingSet
}
