% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\name{identifier-methods}
\alias{identifier-methods}
\alias{identifier}
\alias{identifier,GatingSet-method}
\alias{identifier,GatingSetList-method}
\alias{identifier<-,GatingSetList,ANY-method}
\alias{identifier<-,GatingSet,ANY-method}
\alias{identifier<-,GatingSet-method}
\alias{identifier<-,GatingSet,character-method}
\alias{identifier<-,GatingSetList,character-method}
\title{Retrieve/replace the GUID of a GatingSet or GatingSetList}
\usage{
identifier(object)

\S4method{identifier}{GatingSet,ANY}(object) <- value

\S4method{identifier}{GatingSetList,character}(object) <- value
}
\arguments{
\item{object}{a \code{GatingSet} or \code{GatingSetList}}

\item{value}{string}
}
\description{
Retrieve or replace the GUID (globally unique identifier) for a \code{\link[flowWorkspace:GatingSet-class]{GatingSet}} or \code{\link[flowWorkspace:GatingSetList-class]{GatingSetList}}
}
