\name{keyword}
\alias{keyword}
\alias{keyword,GatingHierarchy,character}
\alias{keyword,GatingHierarchy,character-method}
\alias{keyword,GatingSet,character}
\alias{keyword,GatingSet,character-method}
\title{Methods to retrieve keywords associated with an FCS sample}
\arguments{
  \item{object}{\code{GatingHierarchy) or \code{GatingSet}}

  \item{keyword}{\code{character}.}
}
\description{
  See Methods for function \code{keyword} in Package
  `flowCore'
}
\details{
  Retrieve a specific keyword for a specific sample in a
  \code{GatingHierarchy} or or set of samples in a
  \code{GatingSet}
}
\seealso{
  \code{\link[flowCore]{keyword-methods}}
}

