% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoset.R, R/GatingSet_Methods.R
\docType{methods}
\name{lapply,cytoset-method}
\alias{lapply,cytoset-method}
\alias{lapply,GatingSet-method}
\title{apply \code{FUN} to each sample (i.e. \code{cytoframe})}
\usage{
\S4method{lapply}{cytoset}(X, FUN, ...)

\S4method{lapply}{GatingSet}(X, FUN, ...)
}
\arguments{
\item{X}{\code{cytoset}}

\item{FUN}{\code{function} to be applied to each sample in 'cytoset'}

\item{...}{other arguments to be passed to 'FUN'}

\item{X}{\code{GatingSet}}

\item{FUN}{\code{function} to be applied to each sample in 'GatingSet'}

\item{...}{other arguments to be passed to 'FUN'}
}
\description{
sample names are used for names of the returned list

sample names are used for names of the returned list
}
