% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{lapply-methods}
\alias{lapply-methods}
\alias{lapply}
\alias{lapply,GatingSet-method}
\alias{lapply,cytoset-method}
\title{apply \code{FUN} to each sample (i.e. \code{GatingHierarchy} or \code{cytoframe})
in a \code{GatingSet} or \code{cytoset}}
\usage{
lapply(X, FUN, ...)
}
\arguments{
\item{X}{\code{GatingSet} or \code{cytoset}}

\item{FUN}{\code{function} to be applied to each sample in 'GatingSet' or 'cytoset'}

\item{...}{other arguments to be passed to 'FUN'}
}
\description{
sample names are used for names of the returned list
}
