% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet-Methods.R
\docType{methods}
\name{length,GatingSet-method}
\alias{length}
\alias{length,GatingSet-method}
\alias{show,GatingSet-method}
\title{Methods to get the length of a GatingSet}
\usage{
\S4method{length}{GatingSet}(x)

\S4method{show}{GatingSet}(object)
}
\arguments{
\item{x}{\code{GatingSet}}

\item{object}{\code{object}}
}
\description{
Return the length of a \code{GatingSet} or \code{GatingSetList} object (number of samples).
}

