% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{load_cytoframe}
\alias{load_cytoframe}
\title{Load the cytoframe from disk}
\usage{
load_cytoframe(uri, ...)
}
\arguments{
\item{uri}{path to the cytoframe file}

\item{...}{readonly logical flag indicating whether to open h5 data as readonly. Default is TRUE.
on_disk logical flag indicating whether to keep the data on disk and load it on demand. Default is TRUE.
cred credentials for s3 access. It is a list containing elements of "AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY", "AWS_REGION"
           when NULL, read the default credential file from disk (e.g., ~/.aws/credentials)}
}
\description{
Load the cytoframe from disk
}
