% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{load_cytoframe}
\alias{load_cytoframe}
\title{Load the cytoframe from disk}
\usage{
load_cytoframe(uri, on_disk = TRUE, readonly = on_disk)
}
\arguments{
\item{uri}{path to the cytoframe file}

\item{on_disk}{logical flag indicating whether to keep the data on disk and load it on demand. Default is TRUE.}

\item{readonly}{logical flag indicating whether to open h5 data as readonly. Default is TRUE.
And it is valid when on_disk is set to true.}
}
\description{
Load the cytoframe from disk
}
\seealso{
Other cytoframe/cytoset IO functions: 
\code{\link{cf_get_uri}()},
\code{\link{cf_write_disk}()},
\code{\link{cf_write_h5}()},
\code{\link{cs_get_uri}()},
\code{\link{load_cytoframe_from_fcs}()},
\code{\link{load_cytoset_from_fcs}()}
}
\concept{cytoframe/cytoset IO functions}
