% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cytoframe.R
\name{load_cytoframe_from_h5}
\alias{load_cytoframe_from_h5}
\title{Load the cytoframe from h5 format}
\usage{
load_cytoframe_from_h5(filename, readonly = TRUE, on_disk = TRUE, cred = NULL)
}
\arguments{
\item{filename}{the full path of the output h5 file}

\item{readonly}{logical flag indicating whether to open h5 data as readonly. Default is TRUE.}

\item{on_disk}{logical flag indicating whether to keep the data on disk and load it on demand. Default is TRUE.}

\item{cred}{credentials for s3 access. It is a list containing elements of "AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY", "AWS_REGION"
when NULL, read the default credential file from disk (e.g., ~/.aws/credentials)}
}
\description{
Load the cytoframe from h5 format
}
\seealso{
Other cytoframe/cytoset IO functions: 
\code{\link{cf_get_uri}()},
\code{\link{cf_write_disk}()},
\code{\link{cf_write_h5}()},
\code{\link{cf_write_tile}()},
\code{\link{cs_get_uri}()},
\code{\link{load_cytoframe_from_fcs}()},
\code{\link{load_cytoset_from_fcs}()}
}
\concept{cytoframe/cytoset IO functions}
