% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R,
%   R/GatingSet_Methods.R
\docType{methods}
\name{markernames,GatingHierarchy-method}
\alias{markernames,GatingHierarchy-method}
\alias{markernames<-,GatingHierarchy-method}
\alias{colnames,GatingHierarchy-method}
\alias{colnames<-,GatingHierarchy-method}
\alias{markernames,GatingSet-method}
\alias{markernames<-,GatingSet-method}
\alias{colnames,GatingSet-method}
\alias{colnames<-,GatingSet-method}
\title{Get/set the column(channel) or marker names}
\usage{
\S4method{markernames}{GatingHierarchy}(object)

\S4method{markernames}{GatingHierarchy}(object) <- value

\S4method{colnames}{GatingHierarchy}(x, do.NULL = "missing",
  prefix = "missing")

\S4method{colnames}{GatingHierarchy}(x) <- value

\S4method{markernames}{GatingSet}(object)

\S4method{markernames}{GatingSet}(object) <- value

\S4method{colnames}{GatingSet}(x, do.NULL = "missing", prefix = "missing")

\S4method{colnames}{GatingSet}(x) <- value
}
\arguments{
\item{value}{named character vector for markernames<-, regular character vector for colnames<-.}

\item{x, object}{GatingHierarchy/GatingSet/GatingSetList}

\item{do.NULL, prefix}{not used.}
}
\description{
It simply calls the methods for the underlying flow data (flowSet/ncdfFlowSet/ncdfFlowList).
}
\examples{
\dontrun{

markers.new <- c("CD4", "CD8")
chnls <- c("<B710-A>", "<R780-A>")
names(markers.new) <- chnls
markernames(gs) <- markers.new

chnls <- colnames(gs)
chnls.new <- chnls
chnls.new[c(1,4)] <- c("fsc", "ssc")
colnames(gs) <-  chnls.new
}
}
