% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R
\docType{methods}
\name{isGated}
\alias{isGated}
\alias{isGated,GatingHierarchy,character-method}
\alias{isNegated}
\alias{isNegated,GatingHierarchy,character-method}
\alias{isHidden}
\alias{isHidden,GatingHierarchy,character-method}
\title{The flags of gate nodes
isGated checks if a node is already gated
isNegated checks if a node is negated.
isHidden checks if a node is hidden.}
\usage{
isGated(obj, y, ...)

\S4method{isGated}{GatingHierarchy,character}(obj, y)

isNegated(obj, y, ...)

\S4method{isNegated}{GatingHierarchy,character}(obj, y)

isHidden(obj, y, ...)

\S4method{isHidden}{GatingHierarchy,character}(obj, y)
}
\arguments{
\item{obj}{GatingHierarchy}

\item{y}{node/gating path}

\item{...}{not used}
}
\description{
The flags of gate nodes
isGated checks if a node is already gated
isNegated checks if a node is negated.
isHidden checks if a node is hidden.
}
