% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R
\name{nodeflags}
\alias{nodeflags}
\alias{gh_pop_is_gated}
\alias{isGated}
\alias{isNegated}
\alias{isHidden}
\alias{gh_pop_is_negated}
\alias{gh_pop_is_hidden}
\alias{gh_pop_is_bool_gate}
\title{The flags of gate nodes}
\usage{
gh_pop_is_gated(obj, y)

gh_pop_is_negated(obj, y)

gh_pop_is_hidden(obj, y)

gh_pop_is_bool_gate(obj, y)
}
\arguments{
\item{obj}{GatingHierarchy}

\item{y}{node/gating path}
}
\description{
gh_pop_is_gated checks if a node is already gated.
gh_pop_is_negated checks if a node is negated.
gh_pop_is_hidden checks if a node is hidden.
}
