% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{openWorkspace}
\alias{openWorkspace}
\title{It is now moved along with entire flowJo parser to CytoML package}
\usage{
openWorkspace(file, ...)
}
\arguments{
\item{file}{xml file}

\item{...}{other arguments}
}
\description{
It is now moved along with entire flowJo parser to CytoML package
}
