% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/flowJoWorkspace-Methods.R
\docType{methods}
\name{openWorkspace,character-method}
\alias{closeWorkspace}
\alias{closeWorkspace,flowJoWorkspace-method}
\alias{openWorkspace}
\alias{openWorkspace,character-method}
\title{Open/Close a flowJo workspace}
\usage{
\S4method{openWorkspace}{character}(file, options = 0, ...)

\S4method{closeWorkspace}{flowJoWorkspace}(workspace)
}
\arguments{
\item{file}{Full path to the XML flowJo workspace file.}

\item{options}{xml parsing options passed to \code{\link{xmlTreeParse}}}

\item{...}{other arguments passed to \code{\link{xmlTreeParse}}}

\item{workspace}{A \code{flowJoWorkspace}}
}
\value{
a \code{flowJoWorkspace} object.
}
\description{
Open a flowJo workspace and return a \code{flowJoWorkspace} object.
Close a flowJoWorkspace, destroying the internal representation of the XML document, and freeing the associated memory.
}
\details{
Open an XML flowJo workspace file and return a \code{flowJoWorkspace} object. The workspace is represented using a \code{XMLInternalDocument} object.
	Close a flowJoWorkpsace after finishing with it. This is necessary to explicitly clean up the C-based representation of the XML tree. (See the XML package).
}
\examples{
\dontrun{
	file<-"myworkspace.xml"
	ws<-openWorkspace(file);
	class(ws); #flowJoWorkspace
	closeWorkspace(ws);
}
}

