% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read.gatingML.cytobank.R
\name{parse.gatingML}
\alias{parse.gatingML}
\title{A wrapper that parse the gatingML and FCS files into GatingSet}
\usage{
parse.gatingML(xml, FCS)
}
\arguments{
\item{xml}{the full path of gatingML file}

\item{FCS}{FCS files to be loaded}
}
\value{
a GatingSet
}
\description{
A wrapper that parse the gatingML and FCS files into GatingSet
}
\examples{
\dontrun{
xmlfile <- system.file("extdata/cytotrol_tcell_cytobank.xml", package = "flowWorkspace")
fcsFiles <- list.files(pattern = "CytoTrol", system.file("extdata", package = "flowWorkspaceData"), full = T)
gs <- parse.gatingML(xmlfile, fcsFiles)
plotGate(gs[[1]])
}
}

