% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{pData-methods}
\alias{pData-methods}
\alias{pData}
\alias{pData,GatingHierarchy-method}
\alias{pData,GatingSet-method}
\alias{pData<-,GatingSet,data.frame-method}
\alias{pData<-}
\alias{pData<-,GatingSetList,data.frame-method}
\title{read/set pData of flow data associated with \code{GatingHierarchy}, \code{GatingSet}, or \code{GatingSetList}}
\usage{
pData(object)

pData(object) <- value
}
\arguments{
\item{object}{\code{GatingSet} or \code{GatingSetList}}

\item{value}{\code{data.frame} The replacement of pData for \code{flowSet} or \code{ncdfFlowSet} object}
}
\value{
a \code{data.frame}
}
\description{
Accessor method that gets or replaces the pData of the flowset/ncdfFlowSet object in a GatingHierarchy, GatingSet, or GatingSetList
}
