% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{plotPopCV,GatingHierarchy-method}
\alias{plotPopCV}
\alias{plotPopCV,GatingHierarchy-method}
\alias{plotPopCV,GatingSet-method}
\title{Plot the coefficient of variation between flowJo and flowCore population statistics for each population in a gating hierarchy.}
\usage{
\S4method{plotPopCV}{GatingHierarchy}(x, m = 2, n = 2, ...)

\S4method{plotPopCV}{GatingSet}(x, scales = list(x = list(rot = 90)), ...)
}
\arguments{
\item{x}{A \code{GatingHierarchy} from or a \code{GatingSet}.}

\item{m}{\code{numeric} The number of rows in the panel plot. Now deprecated, uses lattice.}

\item{n}{\code{numeric} The number of columns in the panel plot. Now deprecated, uses lattice.}

\item{scales}{\code{list} see \link{barchart}}

\item{\dots}{Additional arguments to the \code{barplot} methods.}
}
\value{
Nothing is returned.
}
\description{
This function plots the coefficient of variation calculated between the flowJo population statistics and the flowCore population statistics for each population in a gating hierarchy extracted from a flowJoWorkspace.
}
\details{
The CVs are plotted as barplots across panels on a grid of size \code{m} by \code{n}.
}
\examples{
\dontrun{
    #G is a GatingHierarchy
    plotPopCV(G,4,4);
  }
}
\seealso{
\code{\link{getPopStats}}
}

