% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingHierarchy_Methods.R,
%   R/GatingSet_Methods.R
\docType{methods}
\name{plotPopCV,GatingHierarchy-method}
\alias{plotPopCV,GatingHierarchy-method}
\alias{plotPopCV,GatingSet-method}
\alias{plotPopCV}
\title{Plot the coefficient of variation between xml and openCyto population statistics for each population in a gating hierarchy.}
\usage{
\S4method{plotPopCV}{GatingHierarchy}(x, m = 2, n = 2, path = "auto", ...)

\S4method{plotPopCV}{GatingSet}(x, scales = list(x = list(rot = 90)),
  path = "auto", ...)
}
\arguments{
\item{x}{A \code{GatingHierarchy} from or a \code{GatingSet}.}

\item{m}{\code{numeric} The number of rows in the panel plot. Now deprecated, uses lattice.}

\item{n}{\code{numeric} The number of columns in the panel plot. Now deprecated, uses lattice.}

\item{path}{\code{character} see \link{getNodes}}

\item{\dots}{Additional arguments to the \code{barplot} methods.}

\item{scales}{\code{list} see \link{barchart}}
}
\value{
Nothing is returned.
}
\description{
This function plots the coefficient of variation calculated between the xml population statistics and the openCyto population statistics for each population in a gating hierarchy extracted from a xml Workspace.
}
\details{
The CVs are plotted as barplots across panels on a grid of size \code{m} by \code{n}.
}
\examples{
  \dontrun{
    #G is a GatingHierarchy
    plotPopCV(G,4,4);
  }
}
\seealso{
\code{\link{getPopStats}}
}
