% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R
\name{sampleNames}
\alias{sampleNames}
\alias{sampleNames,GatingSet-method}
\alias{sampleNames,cytoset-method}
\alias{sampleNames,cytoset,}
\alias{sampleNames<-,GatingSet,ANY-method}
\alias{sampleNames<-}
\alias{sampleNames<-,GatingSet-method}
\title{Get/update sample names in a GatingSet}
\usage{
sampleNames(object)

sampleNames(object) <- value
}
\arguments{
\item{object}{a \code{GatingSet}}

\item{value}{\code{character} new sample names}
}
\value{
A character vector of sample names
}
\description{
Return  a sample names contained in a GatingSet
}
\details{
The sample names comes from pdata of fs.
}
\examples{
      \dontrun{
        #G is  a GatingSet
        sampleNames(G)
      }
}
