% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingHierarchy-Methods.R, R/GatingSet-Methods.R
\docType{methods}
\name{sampleNames,GatingHierarchy-method}
\alias{sampleNames}
\alias{sampleNames,GatingHierarchy-method}
\alias{sampleNames,GatingSet-method}
\alias{sampleNames<-}
\alias{sampleNames<-,GatingSet,ANY-method}
\alias{sampleNames<-,GatingSet-method}
\title{Get/update sample names in a GatingSet}
\usage{
\S4method{sampleNames}{GatingHierarchy}(object)

\S4method{sampleNames}{GatingSet}(object)

\S4method{sampleNames}{GatingSet}(object) <- value
}
\arguments{
\item{object}{or a \code{GatingSet}}

\item{value}{\code{character} new sample names}
}
\value{
A character vector of sample names
}
\description{
Return  a sample names contained in a GatingSet
}
\details{
The sample names comes from pdata of fs.
}
\examples{
\dontrun{
        #G is  a GatingSet
        sampleNames(G)
      }
}

