% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_gs.R
\name{save_gs}
\alias{save_gs}
\alias{load_gs}
\alias{save_gslist}
\alias{load_gslist}
\title{save/load a GatingSet/GatingSetList to/from disk.}
\usage{
save_gs(
  gs,
  path,
  cdf = c("copy", "move", "skip", "symlink", "link"),
  cred = NULL,
  ...
)

load_gs(
  path,
  h5_readonly = TRUE,
  select = character(),
  verbose = FALSE,
  cred = NULL
)

save_gslist(gslist, path, ...)

load_gslist(path)
}
\arguments{
\item{path}{A character scalar giving the path to save/load the GatingSet to/from.}

\item{cdf}{a character scalar. The valid options are :"copy","move","skip","symlink" specifying what to do with the cdf data file.
Sometimes it is more efficient to move or create a symlink of the existing cdf file to the archived folder.
It is useful to "skip" archiving cdf file if raw data has not been changed.}

\item{cred}{credentials for s3 access. It is a list containing elements of "AWS_ACCESS_KEY_ID", "AWS_SECRET_ACCESS_KEY", "AWS_REGION"
when NULL, read the default credential file from disk (e.g., ~/.aws/credentials)}

\item{...}{other arguments: not used.}

\item{h5_readonly}{whether to open h5 data as read-only. Default is TRUE}

\item{select}{an integer or character vector to select a subset of samples to load}

\item{verbose}{logical flag to optionally print the versions of the libraries that were used to archive the GatingSet for troubleshooting purpose.}

\item{gslist}{A \code{GatingSetList}}

\item{G}{A \code{GatingSet}}
}
\value{
\code{load_gs} returns a GatingSet object
\code{load_gslist} returns a GatingSetList object
}
\description{
Save/load a GatingSet/GatingSetList which is the gated flow data including gates and populations to/from the disk.
The \code{GatingSet} object The internal C data structure (gating tree),\code{ncdfFlowSet} object(if applicable)
}
\examples{
\dontrun{
	#G is a GatingSet
	save_gs(G,path="tempFolder")
	G1<-load_gs(path="tempFolder")

	#G is a GatingSet

	save_gslist(gslist1,path="tempFolder")
	gslist2<-load_gslist(path="tempFolder")
}
}
\seealso{
\code{\link{GatingSet-class}},\code{\link{GatingSetList-class}}
}
