% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GatingSet_Methods.R, R/load_gs.R
\name{save_gs}
\alias{save_gs}
\alias{load_gs}
\alias{save_gslist}
\alias{load_gslist}
\title{save/load a GatingSet/GatingSetList to/from disk.}
\usage{
save_gs(gs, path, cdf = c("copy", "move", "skip", "symlink", "link"),
  ...)

load_gs(path)

save_gs(gs, path, cdf = c("copy", "move", "skip", "symlink", "link"),
  ...)

load_gs(path)

save_gslist(gslist, path, ...)

load_gslist(path)
}
\arguments{
\item{path}{A character scalar giving the path to save/load the GatingSet to/from.}

\item{cdf}{a character scalar. The valid options are :"copy","move","skip","symlink","link" specifying what to do with the cdf data file.
Sometime it is more efficient to move or create a link of the existing cdf file to the archived folder.
It is useful to "skip" archiving cdf file if raw data has not been changed.}

\item{...}{other arguments: not used.}

\item{gslist}{A \code{GatingSetList}}

\item{G}{A \code{GatingSet}}

\item{overwrite}{A logical scalar specifying whether to overwrite the existing folder.}

\item{G}{A \code{GatingSet}}

\item{gslist}{A \code{GatingSetList}}

\item{path}{A character scalar giving the path to save/load the GatingSet to/from.}

\item{cdf}{a character scalar. The valid options are :"copy","move","skip","symlink","link" specifying what to do with the cdf data file.
Sometime it is more efficient to move or create a link of the existing cdf file to the archived folder.
It is useful to "skip" archiving cdf file if raw data has not been changed.}

\item{...}{other arguments: not used.}
}
\value{
\code{load_gs} returns a GatingSet object
\code{load_gslist} returns a GatingSetList object

\code{load_gs} returns a GatingSet object
\code{load_gslist} returns a GatingSetList object
}
\description{
Save/load a GatingSet/GatingSetList which is the gated flow data including gates and populations to/from the disk.
The \code{GatingSet} object The internal C data structure (gating tree),\code{ncdfFlowSet} object(if applicable)

Save/load a GatingSet/GatingSetList which is the gated flow data including gates and populations to/from the disk.
The \code{GatingSet} object The internal C data structure (gating tree),\code{ncdfFlowSet} object(if applicable)
}
\examples{
\dontrun{
	#G is a GatingSet
	save_gs(G,path="tempFolder")
	G1<-load_gs(path="tempFolder")

	#G is a GatingSet

	save_gslist(gslist1,path="tempFolder")
	gslist2<-load_gslist(path="tempFolder")
}
\dontrun{
	#G is a GatingSet
	save_gs(G,path="tempFolder")
	G1<-load_gs(path="tempFolder")

	#G is a GatingSet

	save_gslist(gslist1,path="tempFolder")
	gslist2<-load_gslist(path="tempFolder")
}
}
\seealso{
\code{\link{GatingSet-class}},\code{\link{GatingSetList-class}}

\code{\link{GatingSet-class}},\code{\link{GatingSetList-class}}
}
