% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet-Methods.R
\name{subset.GatingSet}
\alias{subset.GatingSet}
\title{subset the GatingSet/GatingSetList based on 'pData'}
\usage{
\method{subset}{GatingSet}(x, subset, ...)
}
\arguments{
\item{x}{\code{GatingSet} or \code{GatingSetList}}

\item{subset}{logical expression(within the context of pData) indicating samples to keep. see \code{\link[base:subset]{subset}}}

\item{...}{other arguments. (not used)}
}
\value{
a code{GatingSet} or \code{GatingSetList} object
}
\description{
subset the GatingSet/GatingSetList based on 'pData'
}

