% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GatingSet-Methods.R, R/GatingSetList-Methods.R
\docType{methods}
\name{transform,GatingSet-method}
\alias{transform,GatingSet-method}
\alias{transform,GatingSetList-method}
\title{tranform the flow data asssociated with the GatingSet}
\usage{
\S4method{transform}{GatingSet}(`_data`, ...)

\S4method{transform}{GatingSetList}(`_data`, ...)
}
\arguments{
\item{_data}{\code{GatingSet} or \code{GatingSetList}}

\item{...}{expect a \code{transformList} object}
}
\value{
a \code{GatingSet} or \code{GatingSetList} object with the underling flow data transformed.
}
\description{
The transformation functions are saved in the GatingSet and can be retrieved by \link{getTransformations}.
Currently only flowJo-type biexponential transformation(either returned by \link{getTransformations} or constructed by \link{flowJoTrans})
is supported.
}
\examples{
\dontrun{
#construct biexponential transformation function
biexpTrans <- flowJoTrans(channelRange=4096, maxValue=262144, pos=4.5,neg=0, widthBasis=-10)
#make a transformList object
transFuns <- transformList(chnls, biexpTrans)
#add it to GatingSet
gs_trans <- transform(gs, transFuns)
}
}

